refactor(serve): extract frontend assets
