+++ cli/src/main/kotlin/ee/schimke/composeai/cli/serve/ServeWeb.kt
@@ -6,13 +6,9 @@ import kotlinx.serialization.json.JsonObject
 import kotlinx.serialization.json.JsonPrimitive
 
 /**
- * In-code HTML/CSS/JS for the `compose-preview serve` web surface. Generated as Kotlin raw strings
- * (the house style — see [ee.schimke.composeai.cli.WebEmbed]) so the token + preview ids inject
- * straight into the pages with no build step. Two surfaces: a landing page listing the module's
- * previews, and a viewer page with override controls that re-point an `<img>` at `/render`.
- *
- * The viewer is written against a `data-mode` seam (snapshot today) so the future in-browser CMP
- * (`live`) transport mounts into the same page rather than a parallel one.
+ * Server-rendered HTML for the `compose-preview serve` web surface. Large static CSS/JS lives in
+ * classpath assets served by [ServeWebAssets]; this object keeps the dynamic HTML and small
+ * value-injected bootstraps that need token, session, or preview data.
  */
 object ServeWeb {
 
@@ -25,537 +21,24 @@ object ServeWeb {
    */
   data class UnfurlMetadata(val pageUrl: String, val imageUrl: String? = null)
 
-  private val STYLE =
-    """
-    :root { color-scheme: light dark; }
-    * { box-sizing: border-box; }
-    body { margin: 0; padding: 24px; font-family: system-ui, -apple-system, Segoe UI, Roboto, sans-serif;
-      color: #1b1b1f; background: #fafafb; }
-    a { color: #5b5bd6; text-decoration: none; }
-    a:hover { text-decoration: underline; }
-    .cp-head { margin: 0 0 4px; font-size: 1.2rem; font-weight: 600; }
-    .cp-sub { margin: 0 0 20px; font-size: 0.82rem; color: #6b6b70; }
-    /* Per-preview link to the source file on GitHub, under the viewer title. */
-    .cp-source { margin: -12px 0 20px; font-size: 0.8rem; }
-    .cp-source-link { display: inline-flex; align-items: center; gap: 5px; color: #6b6b70;
-      text-decoration: none; }
-    .cp-source-link:hover { text-decoration: underline; }
-    .cp-about { margin: 0 0 24px; padding: 14px 16px; border: 1px solid #e3e3e8; border-radius: 10px;
-      background: #fff; max-width: 720px; }
-    .cp-about-title { margin: 0 0 6px; font-size: 0.95rem; font-weight: 600; }
-    .cp-about-body { margin: 0 0 8px; font-size: 0.84rem; line-height: 1.45; color: #45454c; }
-    .cp-about-body code { font-size: 0.8rem; padding: 0 3px; border-radius: 4px; background: #f0f0f3; }
-    .cp-about-links { margin: 0; font-size: 0.8rem; color: #6b6b70; display: flex; flex-wrap: wrap;
-      align-items: center; gap: 6px; }
-    .cp-about-links a { display: inline-flex; align-items: center; gap: 4px; }
-    .cp-gh { width: 15px; height: 15px; vertical-align: text-bottom; }
-    .cp-about-ver { font-family: ui-monospace, SFMono-Regular, Menlo, monospace; font-size: 0.74rem;
-      padding: 1px 7px; border-radius: 999px; border: 1px solid #d7d7de; background: #fff;
-      color: #45454c; }
-    /* Catalog "back to home" button — replaces the in-catalog design-systems nav row. */
-    .cp-back { display: inline-flex; align-items: center; gap: 6px; margin: 0 0 14px;
-      font-size: 0.82rem; padding: 5px 12px; border-radius: 999px; border: 1px solid #d7d7de;
-      background: #fff; color: #45454c; }
-    .cp-back:hover { background: #f0f0f3; text-decoration: none; }
-    /* Provenance strip: how/when/from-what this catalog was generated (branch, date, versions,
-       a link to re-run the generating workflow). Shown near the catalog header. */
-    .cp-prov { margin: 0 0 18px; padding: 10px 14px; border: 1px solid #e3e3e8; border-radius: 10px;
-      background: #fff; display: flex; flex-wrap: wrap; gap: 6px 16px; font-size: 0.78rem;
-      color: #6b6b70; max-width: 720px; }
-    .cp-prov-item { display: inline-flex; align-items: center; gap: 5px; }
-    .cp-prov-item .cp-prov-key { color: #8a8a92; }
-    .cp-prov-item code { font-size: 0.74rem; padding: 0 4px; border-radius: 4px; background: #f0f0f3; }
-    .cp-systems { margin: 0 0 20px; display: flex; flex-wrap: wrap; align-items: center; gap: 10px;
-      font-size: 0.85rem; }
-    .cp-systems-label { font-weight: 600; color: #6b6b70; }
-    .cp-systems a, .cp-systems-cur { padding: 3px 10px; border-radius: 999px; border: 1px solid #d7d7de; }
-    .cp-systems a { background: #fff; }
-    .cp-systems-cur { background: #ececff; border-color: #c4c4f5; color: #3a3a8a; font-weight: 600; }
-    .cp-toolbar { margin: 0 0 16px; }
-    .cp-searchbar { margin: 0 0 16px; display: flex; align-items: center; gap: 10px; flex-wrap: wrap; }
-    .cp-search { flex: 1 1 260px; max-width: 420px; padding: 7px 12px; font: inherit; font-size: 0.85rem;
-      border: 1px solid #d7d7de; border-radius: 999px; background: #fff; color: inherit; }
-    .cp-search:focus { outline: 2px solid #c4c4f5; outline-offset: 1px; }
-    .cp-count { font-size: 0.78rem; color: #6b6b70; }
-    .cp-empty { margin: 12px 0 0; font-size: 0.85rem; color: #6b6b70; }
-    .cp-empty[hidden] { display: none; }
-    /* Tabbed catalog: a tab bar (role=tablist) over per-section panels. Progressive enhancement —
-       with no JS every section shows under its own <h2> and the tabs act as in-page anchor links;
-       the filter script turns them into real tabs (one panel shown at a time) and adds `cp-js` to
-       <html>, which hides the then-redundant per-section <h2> (the tab already names the section). */
-    .cp-tabs { display: flex; flex-wrap: wrap; gap: 2px 4px; margin: 0 0 18px;
-      border-bottom: 1px solid #e3e3e8; }
-    .cp-tab { display: inline-flex; align-items: center; gap: 7px; padding: 8px 15px;
-      font-size: 0.85rem; color: #6b6b70; border: 1px solid transparent; border-bottom: none;
-      border-radius: 9px 9px 0 0; margin-bottom: -1px; cursor: pointer; }
-    .cp-tab:hover { color: #3a3a8a; background: #f0f0f3; text-decoration: none; }
-    .cp-tab[aria-selected="true"] { color: #3a3a8a; font-weight: 600; background: #fafafb;
-      border-color: #e3e3e8; border-bottom-color: #fafafb; }
-    .cp-tab-count { font-size: 0.7rem; padding: 0 7px; border-radius: 999px; background: #ececff;
-      color: #5a5a9a; }
-    .cp-section { margin: 0; }
-    .cp-section[hidden] { display: none; }
-    .cp-section-head { margin: 0 0 14px; font-size: 1rem; font-weight: 600; }
-    html.cp-js .cp-section-head { display: none; }
-    .cp-subgroup { margin: 0 0 20px; }
-    .cp-subgroup[hidden] { display: none; }
-    .cp-group-head { margin: 18px 0 10px; font-size: 0.72rem; font-weight: 600; letter-spacing: 0.04em;
-      text-transform: uppercase; color: #8a8a92; }
-    .cp-cards { display: grid; grid-template-columns: repeat(auto-fill, minmax(200px, 1fr)); gap: 16px; }
-    /* The Theme picker carries one chip per configured theme (the baked light/dark pair plus any
-       app-declared @ThemeCatalog theme), so it wraps rather than overflowing a narrow viewport. */
-    .cp-theme { display: inline-flex; flex-wrap: wrap; border: 1px solid #d7d7de; border-radius: 999px;
-      overflow: hidden; max-width: 100%; }
-    .cp-theme-btn { border: 0; background: #fff; color: #6b6b70; font: inherit; font-size: 0.78rem;
-      padding: 3px 14px; cursor: pointer; }
-    .cp-theme-btn + .cp-theme-btn { border-left: 1px solid #ececf1; }
-    .cp-theme-btn[aria-pressed="true"] { background: #ececff; color: #3a3a8a; font-weight: 600; }
-    .cp-states { display: inline-flex; flex-wrap: wrap; gap: 6px; margin: 0 0 10px; }
-    .cp-state-btn { border: 1px solid #d7d7de; border-radius: 999px; background: #fff; color: #6b6b70;
-      font: inherit; font-size: 0.78rem; padding: 3px 14px; cursor: pointer; text-decoration: none; }
-    .cp-state-btn:hover { border-color: #b9b9c6; color: #3a3a8a; }
-    .cp-state-btn[aria-current="page"] { background: #ececff; border-color: #c5c5f0; color: #3a3a8a; font-weight: 600; }
-    .cp-badge { display: inline-block; margin-left: 8px; padding: 1px 8px; border-radius: 999px;
-      font-size: 0.7rem; font-weight: 600; vertical-align: middle; white-space: nowrap; }
-    .cp-badge--trusted { background: #e7f4ea; color: #1e7a34; border: 1px solid #b6e0c2; }
-    .cp-badge--unverified { background: #fdf0e3; color: #8a5300; border: 1px solid #f0d3a8; }
-    .cp-degrade { display: flex; flex-wrap: wrap; align-items: baseline; gap: 4px 8px;
-      margin: 0 0 12px; padding: 8px 12px; border-radius: 8px; font-size: 0.8rem; line-height: 1.4;
-      background: #fdf0e3; color: #8a5300; border: 1px solid #f0d3a8; }
-    .cp-degrade-icon { font-weight: 700; }
-    .cp-degrade-item { display: inline; }
-    /* Status page (GET /status): stat tiles, config grid, and the catalog/daemon/failure tables. */
-    .cp-status-grid { display: grid; grid-template-columns: repeat(auto-fill, minmax(210px, 1fr));
-      gap: 10px; margin: 0 0 22px; max-width: 860px; }
-    .cp-stat { border: 1px solid #e3e3e8; border-radius: 10px; background: #fff; padding: 10px 14px; }
-    .cp-stat-key { font-size: 0.71rem; text-transform: uppercase; letter-spacing: 0.04em; color: #8a8a92; }
-    .cp-stat-val { margin-top: 3px; font-size: 1.05rem; font-weight: 600; color: #1b1b1f; }
-    .cp-status-sec { margin: 26px 0 12px; font-size: 1rem; font-weight: 600; }
-    .cp-status-scroll { overflow-x: auto; max-width: 100%; margin: 0 0 8px; }
-    .cp-table { border-collapse: collapse; font-size: 0.82rem; max-width: 860px; min-width: 100%; }
-    .cp-table th { text-align: left; font-weight: 600; color: #6b6b70; font-size: 0.72rem;
-      text-transform: uppercase; letter-spacing: 0.03em; padding: 6px 12px 6px 0;
-      border-bottom: 1px solid #e3e3e8; white-space: nowrap; }
-    .cp-table td { padding: 8px 12px 8px 0; border-bottom: 1px solid #f0f0f3; vertical-align: top; }
-    .cp-table tr:last-child td { border-bottom: none; }
-    .cp-table code { font-size: 0.76rem; padding: 0 4px; border-radius: 4px; background: #f0f0f3; }
-    .cp-ok { color: #1e7a34; font-weight: 600; }
-    .cp-muted { color: #8a8a92; font-size: 0.76rem; }
-    .cp-grid { display: grid; grid-template-columns: repeat(auto-fill, minmax(200px, 1fr)); gap: 16px; }
-    .cp-syslist { grid-template-columns: repeat(auto-fill, minmax(280px, 1fr)); }
-    .cp-syslist .cp-imgwrap { min-height: 120px; }
-    .cp-sys-title { font-size: 0.98rem; font-weight: 600; }
-    .cp-sys-desc { margin-top: 4px; font-size: 0.76rem; color: #6b6b70; line-height: 1.4; }
-    .cp-sys-foot { margin-top: 8px; font-size: 0.74rem; color: #6b6b70; }
-    .cp-sys-noimg { font-size: 0.78rem; color: #a0a0a8; }
-    .cp-card { border: 1px solid #e3e3e8; border-radius: 10px; overflow: hidden; background: #fff;
-      display: block; color: inherit; }
-    .cp-card[hidden] { display: none; }
-    .cp-imgwrap { display: flex; align-items: center; justify-content: center; min-height: 88px;
-      background: #fff; padding: 12px; }
-    /* `width`/`height` stay auto so the max-* pair alone decides the used size and the image can
-       never be stretched. The front door's prebaked heroes DO carry width/height attributes — they
-       give the browser the aspect ratio up front (no layout shift) — but those must not become a
-       definite width, or a max-height clamp (the narrow-viewport rule below drops it to 200px)
-       would squash the height without narrowing the width. The heroes are rasterised at 2x this
-       box, so clamping here is what makes them crisp rather than merely small. */
-    .cp-imgwrap img { max-width: 100%; max-height: 240px; width: auto; height: auto; display: block; }
-    /* A framed thumbnail: a clip window whose aspect-ratio is the component box, that crops away a
-       sticker's empty watch canvas. The window's natural width is the box (so it never upscales past
-       1x), but `max-width: 100%` lets it shrink to a narrow grid card instead of overflowing it —
-       the render <img> inside is absolutely positioned and sized + offset in PERCENTAGES of the box,
-       so the whole frame scales as one when the card is narrower than the box. Overrides fit-to-box. */
-    .cp-crop { position: relative; overflow: hidden; display: block; max-width: 100%; }
-    .cp-imgwrap .cp-crop img { position: absolute; max-width: none; max-height: none; margin: 0; }
-    /* Sticker backing: the preview server shows components on a solid surface by DEFAULT, so a
-       transparent sticker reads like a real component instead of washing out. The header's
-       Background/Transparent toggle flips the whole page to a checkerboard (html.cp-bg-transparent)
-       to inspect the raw transparent PNG; the choice persists per-visitor in localStorage['cp-bg']. */
-    html.cp-bg-transparent .cp-imgwrap, html.cp-bg-transparent .cp-stage {
-      background: repeating-conic-gradient(#f4f4f6 0% 25%, #fff 0% 50%) 50% / 16px 16px; }
-    /* Back each card by its OWN render theme: a light-rendered sticker always sits on a light
-       surface and a dark-rendered one on a dark surface, independent of the browser's color scheme
-       or the explicitly-selected catalog theme. Without this a transparent light sticker (dark
-       strokes) shown on the dark browser backing — or dark on light — washes to near-invisible.
-       Solid mode only; the Transparent toggle's checkerboard still wins via cp-bg-transparent. */
-    html:not(.cp-bg-transparent) .cp-card[data-bg-theme="light"] .cp-imgwrap { background: #fff; }
-    html:not(.cp-bg-transparent) .cp-card[data-bg-theme="dark"] .cp-imgwrap { background: #1d1d20; }
-    /* The single-preview viewer's stage follows the preview's background theme the same way the grid
-       thumbnails do (data-bg-theme on .cp-viewer, kept SEPARATE from the explicit-only
-       data-card-theme filter axis): a dark variant — or any preview in a dark-first system like Wear
-       — sits on a dark stage so a light-on-transparent render stays readable. The viewer JS keeps
-       data-bg-theme in step with the chosen Theme (uiMode) so it doesn't clash after a re-render.
-       Solid mode only; the Transparent checkerboard still wins via cp-bg-transparent. */
-    html:not(.cp-bg-transparent) .cp-viewer[data-bg-theme="dark"] .cp-stage { background: #1d1d20; }
-    html:not(.cp-bg-transparent) .cp-viewer[data-bg-theme="light"] .cp-stage { background: #fff; }
-    .cp-bg { display: inline-flex; border: 1px solid #d7d7de; border-radius: 999px; overflow: hidden; }
-    .cp-bg-btn { border: 0; background: #fff; color: #6b6b70; font: inherit; font-size: 0.78rem;
-      padding: 3px 14px; cursor: pointer; }
-    .cp-bg-btn[aria-pressed="true"] { background: #ececff; color: #3a3a8a; font-weight: 600; }
-    .cp-meta { padding: 8px 10px; font-size: 0.82rem; }
-    .cp-label { font-weight: 500; overflow: hidden; text-overflow: ellipsis; white-space: nowrap; }
-    .cp-id { color: #6b6b70; font-size: 0.7rem; overflow: hidden; text-overflow: ellipsis; white-space: nowrap; }
-    .cp-viewer { display: flex; gap: 24px; flex-wrap: wrap; align-items: flex-start; }
-    .cp-stage { position: relative; flex: 1 1 360px; min-width: 280px; border: 1px solid #e3e3e8;
-      border-radius: 10px;
-      background: #fff; padding: 12px;
-      display: flex; align-items: center; justify-content: center; min-height: 220px; }
-    .cp-stage img, .cp-stage canvas { max-width: 100%; height: auto; }
-    .cp-backend { position: absolute; top: 8px; right: 8px; font-size: 0.62rem; font-weight: 600;
-      letter-spacing: 0.02em; padding: 2px 8px; border-radius: 999px; background: rgba(20,20,22,0.72);
-      color: #fff; pointer-events: none; }
-    .cp-stage iframe { position: absolute; border: 0; background: transparent; opacity: 0;
-      transition: opacity 0.15s ease; pointer-events: none; }
-    .cp-stage iframe.cp-wasm-live { opacity: 1; pointer-events: auto; }
-    /* Live daemon frames paint into the SAME absolute overlay box the Wasm iframe uses, pinned to
-       the snapshot img's slot — so the stage is one fixed box every transport fills and toggling a
-       mode never resizes or shifts it (the img keeps its layout slot underneath via visibility). */
-    .cp-stage canvas.cp-canvas-live { position: absolute; max-width: none; }
-    .cp-live-row { flex-direction: row !important; align-items: center; gap: 6px !important; }
-    .cp-modes { display: flex; flex-direction: row; flex-wrap: wrap; gap: 6px 14px;
-      font-size: 0.8rem; }
-    .cp-controls { flex: 0 0 240px; display: flex; flex-direction: column; gap: 14px; }
-    .cp-controls label { display: flex; flex-direction: column; gap: 4px; font-size: 0.8rem; }
-    .cp-controls input, .cp-controls select { padding: 5px 6px; font-size: 0.85rem; }
-    .cp-status { font-size: 0.75rem; color: #6b6b70; min-height: 1em; }
-    /* Visible mode-activation error overlay: shown when a lane (Live stream, Wasm
-       app, or a /render) fails to activate, so a failed mode reads as an error
-       instead of a silent stale frame. Centered over the stage; [hidden] hides it. */
-    .cp-error { position: absolute; left: 50%; top: 50%; transform: translate(-50%, -50%);
-      max-width: 80%; z-index: 3; padding: 10px 14px; border-radius: 8px;
-      font-size: 0.8rem; line-height: 1.4; text-align: center; color: #8a1c1c;
-      background: #fdecec; border: 1px solid #f3b6b6; box-shadow: 0 1px 4px rgba(0,0,0,0.12); }
-    .cp-error[hidden] { display: none; }
-    .cp-note { font-size: 0.75rem; color: #6b6b70; line-height: 1.4; padding: 8px 10px;
-      border-radius: 8px; background: #f4f4f6; }
-    .cp-controls label:has(:disabled) { opacity: 0.55; }
-    .cp-size { display: flex; flex-direction: column; gap: 8px; }
-    /* :not([hidden]) so the mode toggle's `hidden` attribute still wins over this display rule. */
-    .cp-size-row:not([hidden]) { display: flex; gap: 8px; }
-    .cp-size-row label { flex: 1; }
-    .cp-size-row input { width: 100%; box-sizing: border-box; }
-    .cp-knobs { border-top: 1px solid #e3e3e8; padding-top: 12px; display: flex; flex-direction: column; gap: 8px; }
-    .cp-knobs-head { font-size: 0.72rem; color: #6b6b70; }
-    .cp-overlays { border-top: 1px solid #e3e3e8; padding-top: 12px; display: flex; flex-direction: column; gap: 8px; }
-    .cp-overlays-head { font-size: 0.72rem; color: #6b6b70; }
-    .cp-knobs input:disabled { opacity: 0.7; }
-    .cp-links { border-top: 1px solid #e3e3e8; padding-top: 12px; display: flex; flex-direction: column; gap: 8px; }
-    /* The under-stage column: the lane toggles (Live / Wasm / SVG), the static-snapshot note, and
-       the Export & direct links card. Page-level, not part of the collapsible overrides drawer, so
-       picking a lane and grabbing the /render URLs stay visible with the drawer closed. */
-    .cp-below { margin: 18px 0 0; max-width: 720px; display: flex; flex-direction: column; gap: 12px; }
-    /* Export & direct links: its own border is the card's, so the .cp-links divider rule above is
-       dropped inside it. */
-    .cp-export { padding: 12px 14px; border: 1px solid #e3e3e8;
-      border-radius: 10px; background: #fff; }
-    .cp-export-head { margin: 0 0 10px; font-size: 0.78rem; font-weight: 600; color: #45454c; }
-    .cp-export .cp-links { border-top: 0; padding-top: 0; }
-    .cp-link-row { display: flex; align-items: center; gap: 6px; }
-    .cp-link-kind { font-size: 0.72rem; font-weight: 600; color: #6b6b70; width: 30px; flex: none; }
-    .cp-url { flex: 1; min-width: 0; font-family: ui-monospace, SFMono-Regular, Menlo, monospace;
-      font-size: 0.72rem; padding: 4px 6px; border: 1px solid #d7d7de; border-radius: 6px;
-      background: #fff; color: #1b1b1f; cursor: pointer; }
-    .cp-url.cp-url-copied { border-color: #5b5bd6; box-shadow: 0 0 0 2px rgba(91, 91, 214, 0.25); }
-    .cp-copyimg, .cp-dl { font-size: 0.72rem; padding: 4px 8px; border-radius: 6px; border: 1px solid #d7d7de;
-      background: #fff; color: #5b5bd6; cursor: pointer; text-decoration: none; flex: none; white-space: nowrap; }
-    .cp-copyimg:hover, .cp-dl:hover { background: #f0f0f3; }
-    /* Drawer toggles + the two collapsible drawers (left component nav, right overrides). The open
-       state is a class on .cp-viewer — cp-controls-open (default on) and cp-nav-open (default off) —
-       so a drawer hides purely in CSS and the toggle JS just flips the class + aria-expanded. */
-    .cp-viewer-bar { display: flex; flex-wrap: wrap; gap: 8px; margin: 0 0 14px; }
-    .cp-drawer-toggle { display: inline-flex; align-items: center; gap: 6px; font: inherit;
-      font-size: 0.78rem; padding: 5px 12px; border-radius: 999px; border: 1px solid #d7d7de;
-      background: #fff; color: #45454c; cursor: pointer; }
-    .cp-drawer-toggle:hover { background: #f0f0f3; }
-    .cp-drawer-toggle[aria-expanded="true"] { background: #ececff; border-color: #c4c4f5;
-      color: #3a3a8a; font-weight: 600; }
-    .cp-nav { flex: 0 0 220px; align-self: stretch; max-height: 72vh; overflow: auto;
-      border: 1px solid #e3e3e8; border-radius: 10px; background: #fff; padding: 12px;
-      display: flex; flex-direction: column; gap: 10px; }
-    .cp-nav-head { display: flex; align-items: center; justify-content: space-between;
-      font-size: 0.8rem; font-weight: 600; color: #45454c; }
-    .cp-nav-close { border: 0; background: none; font: inherit; font-size: 1.1rem; line-height: 1;
-      color: #6b6b70; cursor: pointer; padding: 0 2px; }
-    .cp-nav-search { padding: 6px 10px; font: inherit; font-size: 0.8rem; border: 1px solid #d7d7de;
-      border-radius: 8px; background: #fff; color: inherit; }
-    .cp-nav-list { list-style: none; margin: 0; padding: 0; display: flex; flex-direction: column; gap: 2px; }
-    .cp-nav-item { display: flex; align-items: center; gap: 8px; padding: 4px 8px; border-radius: 6px;
-      font-size: 0.8rem; color: inherit; }
-    .cp-nav-thumb { flex: none; width: 28px; height: 28px; object-fit: contain; border-radius: 4px;
-      background: repeating-conic-gradient(#f4f4f6 0% 25%, #fff 0% 50%) 50% / 8px 8px; }
-    .cp-nav-name { overflow: hidden; text-overflow: ellipsis; white-space: nowrap; }
-    .cp-nav-item:hover { background: #f0f0f3; text-decoration: none; }
-    .cp-nav-item[aria-current="page"] { background: #ececff; color: #3a3a8a; font-weight: 600; }
-    .cp-nav-empty { font-size: 0.78rem; color: #6b6b70; }
-    .cp-nav-empty[hidden] { display: none; }
-    .cp-viewer:not(.cp-nav-open) .cp-nav { display: none; }
-    .cp-viewer:not(.cp-controls-open) .cp-controls { display: none; }
-    /* Mobile drawer backdrop — hidden by default; shown (below) only while a drawer is open on a
-       narrow viewport, where the drawers render as bottom sheets. */
-    .cp-scrim { display: none; }
-    /* Collapsible override groups: each axis-set is a <details class="cp-group"> the viewer
-       expands/collapses on its own. The summary is the group title with a rotating disclosure
-       caret; the body holds the labelled controls. Replaces the old flat control stack so the
-       overrides panel reads as a small set of named, foldable groups. */
-    .cp-group { border: 1px solid #e3e3e8; border-radius: 8px; background: #fff; }
-    .cp-group > summary { cursor: pointer; list-style: none; display: flex; align-items: center;
-      gap: 6px; padding: 8px 10px; font-size: 0.78rem; font-weight: 600; color: #45454c;
-      user-select: none; }
-    .cp-group > summary::-webkit-details-marker { display: none; }
-    .cp-group > summary::before { content: "\25B8"; font-size: 0.72rem; color: #8a8a92;
-      transition: transform 0.12s ease; }
-    .cp-group[open] > summary::before { transform: rotate(90deg); }
-    .cp-group > summary:hover { color: #3a3a8a; }
-    .cp-group-body { display: flex; flex-direction: column; gap: 12px; padding: 0 10px 12px; }
-    /* The single Static⇄Live preview toggle (replaces the old PNG/SVG/Live/Wasm radio row). Off =
-       the baked / on-demand snapshot; on = the interactive lane (daemon stream, or the in-browser
-       Wasm app). The hidden mode radios behind it are what the transport JS actually reads. */
-    .cp-preview-mode { display: flex; align-items: center; gap: 8px; flex-wrap: wrap; }
-    .cp-live-toggle { display: inline-flex; align-items: center; gap: 7px; font: inherit;
-      font-size: 0.8rem; font-weight: 600; padding: 6px 14px; border-radius: 999px;
-      border: 1px solid #d7d7de; background: #fff; color: #45454c; cursor: pointer; }
-    .cp-live-toggle:hover:not(:disabled) { background: #f0f0f3; }
-    .cp-live-toggle:disabled { opacity: 0.5; cursor: default; }
-    .cp-live-toggle[aria-pressed="true"] { background: #e7f4ea; border-color: #b6e0c2; color: #1e7a34; }
-    .cp-live-dot { width: 9px; height: 9px; border-radius: 50%; background: #b9b9c6; flex: none; }
-    .cp-live-toggle[aria-pressed="true"] .cp-live-dot { background: #1e9c3f;
-      box-shadow: 0 0 0 3px rgba(30, 156, 63, 0.2); }
-    /* The Remote Compose backend selector: a segmented row of .cp-live-toggle chips (js / java /
-       cmp-android / cmp-jvm) sharing one visual group. */
-    .cp-rc-backends { display: inline-flex; align-items: center; gap: 6px; flex-wrap: wrap; }
-    .cp-rc-backends-label { font-size: 0.7rem; font-weight: 700; letter-spacing: 0.05em;
-      text-transform: uppercase; color: #8a8a92; }
-    .cp-rc-backend { padding: 5px 11px; }
-    .cp-mode-hint { font-size: 0.72rem; color: #6b6b70; }
-    /* The SVG format toggle sits beside the Live toggle (only when the session can export SVG).
-       It swaps the static snapshot between the raster PNG and the resolution-independent vector
-       render; pressing it while Live is on drops back to the static SVG. */
-    .cp-fmt-toggle { display: inline-flex; align-items: center; font: inherit; font-size: 0.78rem;
-      font-weight: 600; padding: 6px 12px; border-radius: 999px; border: 1px solid #d7d7de;
-      background: #fff; color: #45454c; cursor: pointer; }
-    .cp-fmt-toggle:hover { background: #f0f0f3; }
-    .cp-fmt-toggle[aria-pressed="true"] { background: #ececff; border-color: #c4c4f5; color: #3a3a8a; }
-    /* The mode radios are kept in the DOM (the transport JS drives them) but visually removed —
-       the single toggle above is the only visible mode control. */
-    .cp-modes-inputs { position: absolute; width: 1px; height: 1px; overflow: hidden; clip: rect(0 0 0 0); }
-    /* Backend badge flips its accent green when a live lane drives the stage (see backendBadgeScript). */
-    .cp-backend[data-live="true"] { background: rgba(30, 122, 52, 0.82); }
-    /* …and amber while that lane is still activating ("connecting…"), so the wait reads on the
-       preview itself rather than in the controls footer. Declared after the green rule so it wins
-       at equal specificity (a connecting live lane carries both attributes). */
-    .cp-backend[data-pending="true"] { background: rgba(138, 83, 0, 0.86); }
-    /* Shared-document surfaces (POST /docs → GET /d/<id>): the drop zone on the upload page and the
-       played-back document + its facts on the permalink page. Both reuse the card/stage chrome. */
-    .cp-drop { display: flex; flex-direction: column; align-items: center; gap: 8px; max-width: 720px;
-      padding: 28px 20px; border: 2px dashed #d7d7de; border-radius: 12px; background: #fff;
-      text-align: center; cursor: pointer; }
-    .cp-drop:hover, .cp-drop.cp-drop-over { border-color: #8f8ff0; background: #f6f6ff; }
-    .cp-drop-title { font-size: 0.95rem; font-weight: 600; }
-    .cp-drop-hint { font-size: 0.8rem; color: #6b6b70; line-height: 1.45; }
-    .cp-doc-form { max-width: 720px; margin: 16px 0 0; display: flex; flex-wrap: wrap; gap: 8px; }
-    .cp-doc-url { flex: 1 1 320px; padding: 7px 12px; font: inherit; font-size: 0.85rem;
-      border: 1px solid #d7d7de; border-radius: 999px; background: #fff; color: inherit; }
-    .cp-doc-btn { font: inherit; font-size: 0.82rem; font-weight: 600; padding: 6px 16px;
-      border-radius: 999px; border: 1px solid #d7d7de; background: #fff; color: #45454c; cursor: pointer; }
-    .cp-doc-btn:hover { background: #f0f0f3; }
-    .cp-doc-result { max-width: 720px; margin: 18px 0 0; padding: 14px 16px; border-radius: 10px;
-      border: 1px solid #e3e3e8; background: #fff; font-size: 0.84rem; }
-    .cp-doc-result[hidden] { display: none; }
-    .cp-doc-result.cp-doc-error { background: #fdf0e3; color: #8a5300; border-color: #f0d3a8; }
-    .cp-doc-stage { max-width: 720px; border: 1px solid #e3e3e8; border-radius: 10px; background: #fff;
-      padding: 12px; display: flex; align-items: center; justify-content: center; min-height: 260px; }
-    .cp-doc-stage canvas, .cp-doc-stage svg, .cp-doc-stage img { max-width: 100%; height: auto; }
-    .cp-doc-status { margin: 8px 0 0; font-size: 0.78rem; color: #6b6b70; min-height: 1.2em; }
-    .cp-doc-facts { max-width: 720px; margin: 16px 0 0; display: grid;
-      grid-template-columns: repeat(auto-fill, minmax(180px, 1fr)); gap: 10px; }
-    .cp-doc-expiry { display: inline-block; padding: 1px 8px; border-radius: 999px; font-size: 0.7rem;
-      font-weight: 600; background: #fdf0e3; color: #8a5300; border: 1px solid #f0d3a8; }
-    @media (prefers-color-scheme: dark) {
-      body { color: #e6e6e9; background: #161618; }
-      .cp-sub, .cp-id, .cp-status, .cp-about-links, .cp-sys-desc, .cp-sys-foot { color: #a0a0a8; }
-      .cp-card, .cp-stage, .cp-knobs, .cp-links, .cp-about { border-color: #34343a; }
-      .cp-export { background: #1d1d20; border-color: #34343a; }
-      .cp-export-head { color: #c9c9d0; }
-      .cp-url { background: #1d1d20; color: #e6e6e9; border-color: #34343a; }
-      .cp-url.cp-url-copied { border-color: #8f8ff0; box-shadow: 0 0 0 2px rgba(143, 143, 240, 0.3); }
-      .cp-copyimg, .cp-dl { background: #1d1d20; border-color: #34343a; }
-      .cp-copyimg:hover, .cp-dl:hover { background: #26262b; }
-      .cp-card, .cp-about { background: #1d1d20; }
-      .cp-about-body { color: #c9c9d0; }
-      .cp-about-body code { background: #2a2a30; }
-      .cp-systems-label { color: #a0a0a8; }
-      .cp-systems a, .cp-systems-cur { border-color: #34343a; }
-      .cp-systems a { background: #1d1d20; }
-      .cp-systems-cur { background: #26264a; border-color: #45458a; color: #c9c9ff; }
-      .cp-about-ver { background: #1d1d20; border-color: #34343a; color: #c9c9d0; }
-      .cp-back { background: #1d1d20; border-color: #34343a; color: #c9c9d0; }
-      .cp-back:hover { background: #26262b; }
-      .cp-source-link { color: #9a9aa2; }
-      .cp-prov { background: #1d1d20; border-color: #34343a; color: #a0a0a8; }
-      .cp-prov-item .cp-prov-key { color: #7a7a82; }
-      .cp-prov-item code { background: #2a2a30; }
-      .cp-search { border-color: #34343a; background: #1d1d20; }
-      .cp-count, .cp-empty { color: #a0a0a8; }
-      .cp-tabs { border-bottom-color: #34343a; }
-      .cp-tab { color: #a0a0a8; }
-      .cp-tab:hover { color: #c9c9ff; background: #26262b; }
-      .cp-tab[aria-selected="true"] { color: #c9c9ff; background: #161618; border-color: #34343a;
-        border-bottom-color: #161618; }
-      .cp-tab-count { background: #26264a; color: #c9c9ff; }
-      .cp-group-head { color: #7a7a82; }
-      .cp-theme { border-color: #34343a; }
-      .cp-theme-btn { background: #1d1d20; color: #a0a0a8; }
-      .cp-theme-btn + .cp-theme-btn { border-left-color: #34343a; }
-      .cp-theme-btn[aria-pressed="true"] { background: #26264a; color: #c9c9ff; }
-      .cp-state-btn { background: #1d1d20; color: #a0a0a8; border-color: #34343a; }
-      .cp-state-btn:hover { border-color: #4a4a55; color: #c9c9ff; }
-      .cp-state-btn[aria-current="page"] { background: #26264a; border-color: #3a3a6a; color: #c9c9ff; }
-      .cp-note { background: #26262b; color: #a0a0a8; }
-      .cp-drop { background: #1d1d20; border-color: #34343a; }
-      .cp-drop:hover, .cp-drop.cp-drop-over { border-color: #6a6ad0; background: #22222c; }
-      .cp-drop-hint, .cp-doc-status { color: #a0a0a8; }
-      .cp-doc-url, .cp-doc-btn { background: #1d1d20; border-color: #34343a; color: #e6e6e9; }
-      .cp-doc-btn:hover { background: #26262b; }
-      .cp-doc-result, .cp-doc-stage { background: #1d1d20; border-color: #34343a; }
-      /* Stat tiles are shared by the status page and the document facts; without these they stayed
-         white-on-dark in both. */
-      .cp-stat { background: #1d1d20; border-color: #34343a; }
-      .cp-stat-key { color: #7a7a82; }
-      .cp-stat-val { color: #e6e6e9; }
-      .cp-imgwrap, .cp-stage { background: #1d1d20; }
-      html.cp-bg-transparent .cp-imgwrap, html.cp-bg-transparent .cp-stage {
-        background: repeating-conic-gradient(#26262b 0% 25%, #1d1d20 0% 50%) 50% / 16px 16px; }
-      .cp-bg { border-color: #34343a; }
-      .cp-bg-btn { background: #1d1d20; color: #a0a0a8; }
-      .cp-bg-btn[aria-pressed="true"] { background: #26264a; color: #c9c9ff; }
-      .cp-badge--trusted { background: #14361f; color: #6cd98a; border-color: #2c6b40; }
-      .cp-badge--unverified { background: #3a2a12; color: #e6b067; border-color: #6b4f24; }
-      .cp-degrade { background: #3a2a12; color: #e6b067; border-color: #6b4f24; }
-      .cp-drawer-toggle { background: #1d1d20; border-color: #34343a; color: #c9c9d0; }
-      .cp-drawer-toggle:hover { background: #26262b; }
-      .cp-drawer-toggle[aria-expanded="true"] { background: #26264a; border-color: #45458a; color: #c9c9ff; }
-      .cp-nav { background: #1d1d20; border-color: #34343a; }
-      .cp-nav-head { color: #c9c9d0; }
-      .cp-nav-search { background: #1d1d20; border-color: #34343a; }
-      .cp-nav-item:hover { background: #26262b; }
-      .cp-nav-item[aria-current="page"] { background: #26264a; color: #c9c9ff; }
-      .cp-nav-thumb { background: repeating-conic-gradient(#26262b 0% 25%, #1d1d20 0% 50%) 50% / 8px 8px; }
-      .cp-nav-empty { color: #a0a0a8; }
-      .cp-group { background: #1d1d20; border-color: #34343a; }
-      .cp-group > summary { color: #c9c9d0; }
-      .cp-group > summary:hover { color: #c9c9ff; }
-      .cp-live-toggle { background: #1d1d20; border-color: #34343a; color: #c9c9d0; }
-      .cp-live-toggle:hover:not(:disabled) { background: #26262b; }
-      .cp-live-toggle[aria-pressed="true"] { background: #14361f; border-color: #2c6b40; color: #6cd98a; }
-      .cp-fmt-toggle { background: #1d1d20; border-color: #34343a; color: #c9c9d0; }
-      .cp-fmt-toggle:hover { background: #26262b; }
-      .cp-fmt-toggle[aria-pressed="true"] { background: #26264a; border-color: #45458a; color: #c9c9ff; }
-      .cp-mode-hint { color: #a0a0a8; }
-    }
-    /* Mobile / narrow viewports. The viewer is a flex row of stage + overrides (+ optional nav);
-       on a phone we collapse it to a single stacked column and — crucially for usability on a tall
-       preview — turn the two drawers into bottom sheets reachable from a sticky toggle bar, so the
-       overrides and the component list are one tap away instead of a long scroll below the preview.
-       We also drop the flex items' min-width so nothing overflows the viewport, and reflow the
-       grids, size rows and copyable link rows to fit a ~320px screen without horizontal scrolling. */
-    @media (max-width: 640px) {
-      body { padding: 14px; }
-      /* The trust badge can carry a long branch string; constrain it to the viewport and let it
-         wrap instead of forcing a horizontal scroll. */
-      .cp-badge { max-width: 100%; white-space: normal; overflow-wrap: anywhere; }
-      /* Sticky toggle bar so ⚙ Overrides / ☰ Components stay one tap away without scrolling. */
-      .cp-viewer-bar { position: sticky; top: 0; z-index: 20; margin: 0 0 10px; padding: 6px 0;
-        background: #fafafb; }
-      .cp-viewer { gap: 16px; }
-      /* Stage, overrides and nav each stack full-width; min-width:0 lets the flex items (and their
-         selects/inputs) shrink below their content's intrinsic width instead of overflowing. */
-      .cp-stage, .cp-controls, .cp-nav { flex: 1 1 100%; min-width: 0; }
-      .cp-controls label, .cp-group, .cp-group-body { min-width: 0; }
-      .cp-controls select, .cp-controls input { max-width: 100%; }
-      /* Open drawers overlay the preview as bottom sheets (with the scrim below), so overrides and
-         the component list are reachable without scrolling past a tall preview. */
-      .cp-viewer.cp-controls-open .cp-controls,
-      .cp-viewer.cp-nav-open .cp-nav {
-        position: fixed; left: 0; right: 0; bottom: 0; z-index: 40; margin: 0;
-        max-height: 78vh; overflow: auto;
-        padding: 12px 14px calc(14px + env(safe-area-inset-bottom, 0px));
-        border: 0; border-top: 1px solid #e3e3e8; border-radius: 14px 14px 0 0;
-        background: #fafafb; box-shadow: 0 -8px 24px rgba(0, 0, 0, 0.18); }
-      /* Grabber affordance at the top of a sheet. */
-      .cp-viewer.cp-controls-open .cp-controls::before,
-      .cp-viewer.cp-nav-open .cp-nav::before {
-        content: ""; display: block; width: 36px; height: 4px; margin: 0 auto 10px;
-        border-radius: 999px; background: #c9c9d0; }
-      .cp-scrim.cp-scrim-on { display: block; position: fixed; inset: 0; z-index: 35;
-        background: rgba(0, 0, 0, 0.38); }
-      .cp-grid, .cp-cards { grid-template-columns: repeat(auto-fill, minmax(140px, 1fr)); gap: 12px; }
-      .cp-syslist, .cp-syslist.cp-grid { grid-template-columns: 1fr; }
-      .cp-imgwrap img { max-height: 200px; }
-      /* Copyable direct-link rows: the URL field takes its own line above the Copy/Download
-         buttons instead of squeezing all three onto one narrow row. */
-      .cp-link-row { flex-wrap: wrap; }
-      .cp-url { flex: 1 1 100%; }
-      /* The paired Min/Max (and Fixed W/H) size inputs stack rather than sit side by side. */
-      .cp-size-row:not([hidden]) { flex-direction: column; gap: 6px; }
-    }
-    /* Dark-mode surfaces for the mobile sticky bar + bottom sheets. */
-    @media (max-width: 640px) and (prefers-color-scheme: dark) {
-      .cp-viewer-bar { background: #161618; }
-      .cp-viewer.cp-controls-open .cp-controls,
-      .cp-viewer.cp-nav-open .cp-nav { background: #161618; border-top-color: #34343a;
-        box-shadow: 0 -8px 24px rgba(0, 0, 0, 0.5); }
-      .cp-viewer.cp-controls-open .cp-controls::before,
-      .cp-viewer.cp-nav-open .cp-nav::before { background: #4a4a55; }
-    }
-    """
-      .trimIndent()
+  /** In-memory engagement metrics surfaced by the live server UI/API. */
+  data class PreviewEngagement(val views: Long = 0)
 
-  /**
-   * Scoped styles for the [playgroundPage] editor — kept off the global [STYLE] since one page uses
-   * them.
-   */
-  private val PLAYGROUND_STYLE =
-    """
-    .cp-pg { display: flex; flex-direction: column; gap: 12px; margin-top: 16px; }
-    .cp-pg-bar { display: flex; align-items: center; gap: 10px; }
-    .cp-pg-modelabel { font-weight: 600; }
-    .cp-pg-mode { padding: 6px 8px; border-radius: 8px; border: 1px solid rgba(0,0,0,0.18);
-      background: #fff; font: inherit; }
-    .cp-pg-run { margin-left: auto; }
-    .cp-pg-files { display: flex; align-items: center; gap: 8px; flex-wrap: wrap; }
-    .cp-pg-file { padding: 5px 10px; border-radius: 8px; border: 1px solid rgba(0,0,0,0.18);
-      background: #fff; font: inherit; font-size: 13px; cursor: pointer; }
-    .cp-pg-file[aria-current="true"] { background: #1b1b1f; color: #fff; border-color: #1b1b1f; }
-    .cp-pg-filebtn { padding: 5px 10px; border-radius: 8px; border: 1px dashed rgba(0,0,0,0.28);
-      background: transparent; font: inherit; font-size: 13px; cursor: pointer; }
-    .cp-pg-source { width: 100%; min-height: 260px; box-sizing: border-box; padding: 12px 14px;
-      border-radius: 10px; border: 1px solid rgba(0,0,0,0.18); background: #fbfbfd;
-      font-family: ui-monospace, SFMono-Regular, Menlo, Consolas, monospace; font-size: 13px;
-      line-height: 1.5; tab-size: 2; resize: vertical; }
-    .cp-pg-status { font-size: 14px; color: #55555f; }
-    .cp-pg-status.cp-doc-error { color: #b00020; }
-    .cp-pg-diags { list-style: none; margin: 0; padding: 0; display: flex; flex-direction: column;
-      gap: 6px; }
-    .cp-pg-diag { font-family: ui-monospace, SFMono-Regular, Menlo, Consolas, monospace;
-      font-size: 12.5px; padding: 6px 10px; border-radius: 8px; background: #f2f2f5; }
-    .cp-pg-error { background: #fdecef; color: #8a0017; }
-    .cp-pg-warning { background: #fff5e0; color: #7a5200; }
-    .cp-pg-result { display: flex; flex-direction: column; gap: 12px; align-items: flex-start; }
-    .cp-pg-image { max-width: 100%; border-radius: 10px; border: 1px solid rgba(0,0,0,0.12);
-      background: #fff; }
-    @media (prefers-color-scheme: dark) {
-      .cp-pg-mode, .cp-pg-source, .cp-pg-file { background: #161618; border-color: #34343a;
-        color: #e6e6ea; }
-      .cp-pg-file[aria-current="true"] { background: #e6e6ea; color: #161618; border-color: #e6e6ea; }
-      .cp-pg-filebtn { border-color: #4a4a52; color: #e6e6ea; }
-      .cp-pg-status { color: #a8a8b2; }
-      .cp-pg-diag { background: #1d1d20; }
-      .cp-pg-error { background: #34161b; color: #ff9aa6; }
-      .cp-pg-warning { background: #2c2410; color: #ffd486; }
-      .cp-pg-image { border-color: #34343a; }
-    }
-    """
-      .trimIndent()
+  private const val ASSET_BASE = "/assets/serve"
+
+  private fun assetHref(name: String): String = "$ASSET_BASE/$name"
+
+  private fun scriptTag(name: String): String = "<script src=\"${assetHref(name)}\"></script>"
+
+  private fun viewCountHtml(views: Long): String =
+    if (views <= 0) "" else "<div class=\"cp-engage\">${formatCount(views)} views</div>"
+
+  private fun viewerViewCountHtml(views: Long): String =
+    if (views <= 0) "" else "<p class=\"cp-viewer-engage\">${formatCount(views)} views</p>"
+
+  private fun formatCount(n: Long): String =
+    if (n < 1000) n.toString()
+    else String.format(java.util.Locale.ROOT, "%.1f", n / 1000.0).removeSuffix(".0") + "k"
 
   /**
    * The starter snippet the [playgroundPage] editor opens with — a minimal Material 3 `@Preview`.
@@ -1652,7 +1135,7 @@ object ServeWeb {
       // (id carries no __light/__dark segment), which would otherwise open at (default). A themed
       // variant's theme is explicit in its deep link, so it opens on its baked (instant) pixels and
       // is not overridden by a remembered theme — seeding it would force a daemon re-render on a
-      // fresh deep link, defeating baked-until-changed. Runs before viewerScript()'s initial render.
+      // fresh deep link, defeating baked-until-changed. Runs before viewer.js' initial render.
       var root = document.querySelector(".cp-viewer");
       var pid = (root && root.getAttribute("data-preview-id")) || "";
       var themed = pid.split("__").some(function (s) { return s === "light" || s === "dark"; });
@@ -1686,51 +1169,6 @@ object ServeWeb {
         syncBg();
       });
       syncBg();
-    })();
-    """
-      .trimIndent()
-
-  /**
-   * Backend-provenance badge: a small corner label naming the tier that produced the pixels now on
-   * the stage, so a viewer can tell an in-browser CMP render from a baked snapshot. Reads the
-   * viewer's `data-mode` (kept in sync by the transport toggles) — `CMP-WASM` for the Wasm app
-   * (always that), `SVG` for the vector snapshot lane, else the server-supplied `data-live-backend`
-   * / `data-snapshot-backend` label, so the daemon's actual platform (desktop/JVM or Android) and
-   * the snapshot renderer stay accurate.
-   *
-   * The badge doubles as the **activation** indicator: while the viewer carries `data-pending` (set
-   * by [viewerScript] when a live lane is coming up), the badge shows that label instead — `◌
-   * connecting…` on an amber accent — so the wait is visible on the preview rather than as small
-   * grey text under the controls, where it was easy to miss.
-   */
-  private fun backendBadgeScript(): String =
-    """
-    (function () {
-      var root = document.querySelector(".cp-viewer");
-      var badge = document.getElementById("cp-backend");
-      if (!root || !badge) return;
-      // The badge carries an icon that flips with the lane: ▶ for an interactive live lane (the
-      // daemon stream or the in-browser Wasm app), ▪ for the static snapshot — the visible signal
-      // that the Static⇄Live toggle changed state (paired with a green accent via data-live).
-      function isLive(mode) { return mode === "wasm" || mode === "live"; }
-      function label(mode) {
-        if (mode === "wasm") return "▶ CMP-WASM";
-        if (mode === "live") return "▶ " + (root.getAttribute("data-live-backend") || "Live");
-        if (mode === "svg") return "▪ SVG";
-        return "▪ " + (root.getAttribute("data-snapshot-backend") || "Snapshot");
-      }
-      function refresh() {
-        var mode = root.getAttribute("data-mode");
-        // ◌ (an open circle) reads as "not yet painting", distinct from the ▶/▪ lane icons.
-        var pending = root.getAttribute("data-pending");
-        badge.textContent = pending ? "◌ " + pending : label(mode);
-        badge.setAttribute("data-live", isLive(mode) ? "true" : "false");
-        if (pending) badge.setAttribute("data-pending", "true");
-        else badge.removeAttribute("data-pending");
-      }
-      new MutationObserver(refresh)
-        .observe(root, { attributes: true, attributeFilter: ["data-mode", "data-pending"] });
-      refresh();
     })();
     """
       .trimIndent()
@@ -2085,7 +1523,7 @@ object ServeWeb {
       unfurl = unfurl,
       body =
         """
-        <style>$PLAYGROUND_STYLE</style>
+        <link rel="stylesheet" href="${assetHref("playground.css")}">
         <h1 class="cp-head">Playground</h1>
         <p class="cp-sub">Write a Compose snippet, compile it against the live catalog, and open a
           preview. This lane runs your code on the server, so it stays behind your token.</p>
@@ -2999,6 +2437,11 @@ object ServeWeb {
      * card can. Defaults to `{ false }`: a plain static bundle offers baked light/dark only.
      */
     canRenderThemeFor: (String) -> Boolean = { false },
+    /**
+     * Per-preview engagement counts for this running server. The map is additive UI/API metadata:
+     * missing or zero entries render no badge.
+     */
+    engagement: Map<String, PreviewEngagement> = emptyMap(),
     /** Absolute page + representative preview URLs for Open Graph/Twitter link previews. */
     unfurl: UnfurlMetadata? = null,
   ): String {
@@ -3032,6 +2475,8 @@ object ServeWeb {
     // its baked pixels (a themed render would ignore the theme anyway).
     fun themeBase(card: GridCard) =
       renderedVariant(card).let { if (themeRenderable(it)) renderSrc(it) else "" }
+    fun cardViews(card: GridCard): Long =
+      listOfNotNull(card.light, card.dark, card.neutral).sumOf { engagement[it.id]?.views ?: 0L }
     fun swapCard(card: GridCard): String {
       val l = card.light!!
       val d = card.dark!!
@@ -3056,6 +2501,7 @@ object ServeWeb {
           <div class="cp-meta">
             <div class="cp-label" title="${WebEscaping.htmlEscape(def.label)}">${WebEscaping.htmlEscape(def.label)}</div>
             <div class="cp-id">${WebEscaping.htmlEscape(def.id)}</div>
+            ${viewCountHtml(cardViews(card))}
           </div>
         </a>
         """
@@ -3075,6 +2521,7 @@ object ServeWeb {
             <div class="cp-meta">
               <div class="cp-label" title="$idText">$label</div>
               <div class="cp-id">$idText</div>
+              ${viewCountHtml(engagement[p.id]?.views ?: 0L)}
             </div>
           </a>
           """
@@ -3356,6 +2803,8 @@ object ServeWeb {
      * [degradeBanner].
      */
     degradations: List<ServeDegradation> = emptyList(),
+    /** Engagement count for this preview on the running server. */
+    engagement: PreviewEngagement = PreviewEngagement(),
     /** Absolute viewer + PNG URLs for Open Graph/Twitter link previews. */
     unfurl: UnfurlMetadata? = null,
     /**
@@ -3692,6 +3141,7 @@ object ServeWeb {
       """
       <h1 class="cp-head"><a href="$basePath/$q">← previews</a>${trustBadge(trust)}</h1>
       ${degradeBanner(degradations)}<p class="cp-sub" title="$idText">$label</p>${sourceLinkHtml(sourceHref, preview.sourceFile)}
+      ${viewerViewCountHtml(engagement.views)}
       $switchers
       <div class="cp-viewer-bar">
         $navToggle
@@ -3839,11 +3289,11 @@ object ServeWeb {
       <!-- Backdrop shown behind an open drawer on mobile (drawers become bottom sheets there);
            tapping it dismisses the sheet. Inert on desktop. -->
       <div class="cp-scrim" id="cp-scrim" aria-hidden="true"></div>
-      <script>${groupStickyScript()}</script>
-      <script>${drawerScript()}</script>
+      ${scriptTag("viewer-groups.js")}
+      ${scriptTag("viewer-drawers.js")}
       <script>${viewerThemeStickyScript(themeStorageKey(sessionId, basePath))}</script>
-      <script>${viewerScript()}</script>
-      <script>${backendBadgeScript()}</script>
+      ${scriptTag("viewer.js")}
+      ${scriptTag("backend-badge.js")}
       """
         .trimIndent()
     return document(
@@ -3854,1416 +3304,6 @@ object ServeWeb {
       unfurl = unfurl,
     )
   }
-
-  /**
-   * Viewer JS. Two transports behind one set of controls (the `data-mode` seam):
-   * - **snapshot** (default): rebuild the `/render` URL from the controls and swap `img.src`.
-   * - **live** (tier-2): when "Live (stream)" is on, open the `/ws/{id}` WebSocket, paint pushed
-   *   frames onto a `<canvas>`, send `setOverrides` as controls change, and forward pointer drags,
-   *   wheel scroll, and keyboard into the held composition as `input` messages.
-   *
-   * Override collection ([overrides]) is shared by both, so an opt-in field stays opt-in either way
-   * (notably fontScale, which is only sent once the slider is moved).
-   */
-  private fun viewerScript(): String =
-    """
-    (function () {
-      "use strict";
-      var root = document.querySelector(".cp-viewer");
-      var img = document.getElementById("cp-img");
-      var canvas = document.getElementById("cp-canvas");
-      var status = document.getElementById("cp-status");
-      var errorBox = document.getElementById("cp-error");
-      var live = document.getElementById("cp-live");
-      // Tall previews used to size the stage from their full width-constrained height, which could
-      // push the rest of the viewer several screens below the fold. Default to a viewport-bounded
-      // contain fit; "Fit width" deliberately restores the old unconstrained-height presentation.
-      // The snapshot remains the geometry source for Live/Wasm, so re-pin an active overlay after
-      // changing modes.
-      var zoomBtns = document.querySelectorAll(".cp-zoom-btn");
-      function applyZoom(mode) {
-        if (mode !== "width") mode = "fit";
-        var maxHeight = mode === "fit" ? "72vh" : "";
-        img.style.maxHeight = maxHeight;
-        var rcZoomCanvas = document.getElementById("cp-rc-canvas");
-        if (rcZoomCanvas) rcZoomCanvas.style.maxHeight = maxHeight;
-        root.setAttribute("data-zoom", mode);
-        zoomBtns.forEach(function (b) {
-          b.setAttribute(
-            "aria-pressed",
-            b.getAttribute("data-zoom-mode") === mode ? "true" : "false"
-          );
-        });
-        window.requestAnimationFrame(function () {
-          if (live && live.checked && !canvas.hidden) fitLiveCanvas();
-          if (wasmActive()) positionWasmFrame();
-        });
-      }
-      zoomBtns.forEach(function (b) {
-        b.addEventListener("click", function () {
-          applyZoom(b.getAttribute("data-zoom-mode"));
-        });
-      });
-      applyZoom("fit");
-      // Surface a mode-activation failure visibly, instead of leaving a stale frame that reads as a
-      // (wrong) render. Every lane routes its failure here — a dead Live stream, a Wasm app that
-      // never boots, a /render that errors — so "can't activate this mode" is never silent.
-      // Activation state for a lane that hasn't painted yet, surfaced on the stage's backend badge
-      // (see backendBadgeScript) instead of the controls footer — a "connecting…" nobody scrolls to
-      // isn't feedback. Pass null to clear.
-      function setPending(label) {
-        if (label) root.setAttribute("data-pending", label);
-        else root.removeAttribute("data-pending");
-      }
-      function showModeError(msg) {
-        setPending(null);
-        if (!errorBox) { status.textContent = msg; return; }
-        errorBox.textContent = msg;
-        errorBox.hidden = false;
-        status.textContent = "";
-      }
-      function clearModeError() { if (errorBox) { errorBox.hidden = true; errorBox.textContent = ""; } }
-      // Human-readable reason for a Live stream that closed before delivering a frame. Maps the
-      // server's close codes (1013 capacity, 1008 unauthorized, 1003/CANNOT_ACCEPT carries a reason);
-      // a bare abnormal close (1006, e.g. a proxy 502 on the WS upgrade) gets the generic message.
-      function liveCloseReason(ev) {
-        if (ev && ev.code === 1013) return "Live preview is at capacity — try again shortly.";
-        if (ev && ev.code === 1008) return "Live preview unauthorized.";
-        if (ev && ev.reason) return "Live preview unavailable: " + ev.reason;
-        return "Live preview couldn't connect — the live stream may be unavailable on this server.";
-      }
-      // Whether the snapshot lane is static (baked PNGs, no /render re-render) — the explicit signal
-      // for the wasm auto-enable below. NOT `live.disabled`: a trusted-catalog live session serves
-      // static snapshots yet leaves the Live toggle enabled, so `live.disabled` no longer implies
-      // "static".
-      var staticSnapshot = root.getAttribute("data-static-snapshot") === "true";
-      // Whether an override-bearing /render returns fresh pixels even on a static snapshot lane (a
-      // trusted-catalog live session: its carried daemon re-renders author-declared knob edits on
-      // demand). When true, a knob edit re-points the snapshot /render URL rather than sitting dead.
-      var canRenderOverrides = root.getAttribute("data-can-render-overrides") === "true";
-      var previewId = root.getAttribute("data-preview-id");
-      // The session path prefix ("/<system>") when this viewer is served under a path — it sits at
-      // "<base>/p/<id>", so stripping the trailing "/p/<id>" recovers the base ("" for the root
-      // mount / legacy ?session= form). /render + /ws requests are prefixed with it so they hit the
-      // same session without needing ?session= threaded through.
-      var base = location.pathname.replace(/\/p\/[^/]*\/?$/, "");
-      var token = new URLSearchParams(location.search).get("token") || "";
-      // Carry the tenant through follow-up requests so a non-default ?session= stays on its module.
-      var session = new URLSearchParams(location.search).get("session") || "";
-      // Hydrate the declared knob controls from the page URL's `knob.<key>` params, so a deep link
-      // (or a copied "Direct links — overrides applied" URL) opens with those values already set.
-      // The initial render then carries them through whichever transport is live — including the
-      // Wasm iframe, whose fragment/patch is built purely from control state — instead of showing
-      // the author default until the control is manually edited. Only the author-declared knobs this
-      // feature drives are hydrated (display axes are unchanged, pre-existing behaviour).
-      (function () {
-        var q = new URLSearchParams(location.search);
-        document.querySelectorAll(".cp-knob").forEach(function (el) {
-          var key = el.getAttribute("data-knob-key");
-          if (!key) return;
-          var v = q.get("knob." + key);
-          if (v === null) return;
-          if (el.type === "checkbox") el.checked = (v === "true" || v === "1");
-          else el.value = v;
-        });
-        // Remote Compose knobs hydrate from `rc.<name>`, whose value is `<kind>:<value>`; strip the
-        // control's own `data-rc-kind` prefix back off before assigning the bare value.
-        document.querySelectorAll(".cp-rc-knob").forEach(function (el) {
-          var name = el.getAttribute("data-rc-name");
-          if (!name) return;
-          var v = q.get("rc." + name);
-          if (v === null) return;
-          var kind = el.getAttribute("data-rc-kind") || "";
-          if (kind && v.indexOf(kind + ":") === 0) v = v.substring(kind.length + 1);
-          if (el.type === "checkbox") el.checked = (v === "true" || v === "1");
-          else el.value = v;
-        });
-      })();
-      // The selects + text input are opt-in (empty value = "use the preview's default"). The font
-      // scale slider has no empty state, so it's gated separately: we only send fontScale once the
-      // user moves it (fontScaleTouched), otherwise the slider's standing 1.0 would override a
-      // preview's declared default font scale and the first render wouldn't match the thumbnail.
-      var fields = ["uiMode", "device", "localeTag", "orientation", "background"];
-      var fs = document.getElementById("cp-fontScale");
-      var fsVal = document.getElementById("cp-fontScale-val");
-      var fontScaleTouched = false;
-      var ws = null;
-      // The snapshot lane serves either the raster PNG or the vector SVG through the same <img>.
-      // The render-mode radio flips this (".png" default, ".svg" in SVG mode); refreshSnapshot and
-      // the copyable links read it so a re-render / copied URL matches the on-screen format.
-      var snapshotExt = ".png";
-
-      // Size overrides (the Fixed / Max / Min / Within modes). Which query params carry the numbers
-      // is chosen by the mode: Fixed pins the frame via widthPx/heightPx; Max / Min / Within are
-      // wrapped-axis bounds (maxWidthPx / minWidthPx …). Blank inputs are omitted, so one axis can be
-      // bounded without the other. Server-side only (a daemon re-measures) — the inputs are
-      // disabled on a static snapshot like Device/Orientation, so it never emits them.
-      //
-      // The inputs are authored in dp (the Compose unit); the wire stays in px like every other
-      // override, so a dp value is multiplied by the backend's render density before it's sent (and
-      // the copyable /render URL stays px-consistent). data-render-density carries the factor.
-      var renderDensity = parseFloat(root.getAttribute("data-render-density")) || 2;
-      // dp (string from the input) → a positive integer px value, or null when blank/non-positive.
-      function sizePx(id) {
-        var el = document.getElementById(id);
-        if (!el || !el.value) return null;
-        var dp = parseFloat(el.value);
-        if (!(dp > 0)) return null;
-        return String(Math.max(1, Math.round(dp * renderDensity)));
-      }
-      function sizeOverrides() {
-        var mode = document.getElementById("cp-sizeMode");
-        var o = {};
-        if (!mode || !mode.value) return o;
-        if (mode.value === "fixed") {
-          if (sizePx("cp-fixedW")) o.widthPx = sizePx("cp-fixedW");
-          if (sizePx("cp-fixedH")) o.heightPx = sizePx("cp-fixedH");
-        }
-        if (mode.value === "min" || mode.value === "within") {
-          if (sizePx("cp-minW")) o.minWidthPx = sizePx("cp-minW");
-          if (sizePx("cp-minH")) o.minHeightPx = sizePx("cp-minH");
-        }
-        if (mode.value === "max" || mode.value === "within") {
-          if (sizePx("cp-maxW")) o.maxWidthPx = sizePx("cp-maxW");
-          if (sizePx("cp-maxH")) o.maxHeightPx = sizePx("cp-maxH");
-        }
-        return o;
-      }
-      function overrides() {
-        var o = {};
-        fields.forEach(function (f) {
-          var el = document.getElementById("cp-" + f);
-          if (el && !el.disabled && el.value) o[f] = el.value;
-        });
-        if (fontScaleTouched && fs) o.fontScale = fs.value;
-        var size = sizeOverrides();
-        Object.keys(size).forEach(function (k) { o[k] = size[k]; });
-        return o;
-      }
-      // The live-stream override map: the display fields PLUS the author-declared knob values as
-      // `knob.<key>=<value>` entries (the daemon's setOverrides parses the same map /render does,
-      // typing each from the preview's declaration). Kept separate from overrides() so query() and
-      // the Wasm patch — which append/ignore knobs their own way — are unaffected; without this a
-      // knob edit during an active Live (stream) would send only the display fields and the daemon
-      // would reset the others to their defaults. Unlike query(), every knob is sent (not just
-      // changed ones) for exactly that reason, so defaults are not filtered here.
-      function liveOverrides() {
-        var o = overrides();
-        document.querySelectorAll(".cp-knob").forEach(function (el) {
-          if (el.disabled) return;
-          var key = el.getAttribute("data-knob-key");
-          if (!key) return;
-          var val = (el.type === "checkbox") ? (el.checked ? "true" : "false") : el.value;
-          if (val === "") return;
-          o["knob." + key] = val;
-        });
-        // Remote Compose knobs carry their own `<kind>:` tag: `rc.<name>=<kind>:<value>`. Sent for
-        // every RC knob (like the plain knobs) so a Live setOverrides doesn't reset the others.
-        document.querySelectorAll(".cp-rc-knob").forEach(function (el) {
-          if (el.disabled) return;
-          var name = el.getAttribute("data-rc-name");
-          if (!name) return;
-          var kind = el.getAttribute("data-rc-kind") || "string";
-          var val = (el.type === "checkbox") ? (el.checked ? "true" : "false") : el.value;
-          if (val === "") return;
-          o["rc." + name] = kind + ":" + val;
-        });
-        // Live-only overlay toggles (talkBack / touchOverlay). Their id is "cp-<key>", so the daemon
-        // key is the id minus the prefix. Sent as an explicit true/false so unchecking clears the
-        // overlay on the next setOverrides (which replaces the whole map). Disabled ⇒ not live ⇒
-        // skipped, so they never leak onto a snapshot.
-        document.querySelectorAll(".cp-overlay").forEach(function (el) {
-          if (el.disabled) return;
-          o[el.id.replace(/^cp-/, "")] = el.checked ? "true" : "false";
-        });
-        // App-declared theme (themeProvider = provider FQN). Only when a theme is picked and the
-        // control is live; "(default)" (empty) leaves the daemon on the preview's own wrapper.
-        var tp = document.getElementById("cp-themeProvider");
-        if (tp && !tp.disabled && tp.value) o["themeProvider"] = tp.value;
-        // Detected-feature: keyboard focus. Checked ⇒ focus the first focusable + draw the overlay
-        // (focus=0). Daemon-only, so skipped when disabled.
-        var fc = document.getElementById("cp-focus");
-        if (fc && !fc.disabled && fc.checked) o["focus"] = "0";
-        // Detected-feature: one-handed gesture hints. Checked ⇒ draw the gesture-hint overlay
-        // (gestures=true). Android-daemon-only, so skipped when disabled.
-        var gc = document.getElementById("cp-gestures");
-        if (gc && !gc.disabled && gc.checked) o["gestures"] = "true";
-        return o;
-      }
-      // RC backend selector state (the #cp-rc-backends chips). `rcPlayerBackend` is the current pick
-      // and `rcPlayerPicked` gates whether it rides the render URL — so page load stays on the
-      // instant default snapshot until the visitor actually chooses a server-side backend.
-      var rcBackendsEl = document.getElementById("cp-rc-backends");
-      var rcPlayerBackend = rcBackendsEl ? (rcBackendsEl.getAttribute("data-default") || "") : "";
-      var rcPlayerPicked = false;
-      // Reconcile the backend chips' pressed state with the active lane. Hoisted (the real impl is
-      // assigned in the selector block below) so the common mode-transition path (enterMode) can
-      // call it whenever the viewer leaves a lane through ANY control — not only a chip click — so
-      // the JS chip can't stay pressed after e.g. Live/Wasm/SVG takes over the stage. A no-op stub
-      // for a non-Remote-Compose preview (no selector present).
-      var syncRcBackendChips = function () {};
-      function query() {
-        var o = overrides();
-        // Public routes are open, so a page that arrived without a token stays token-free — only
-        // carry token= when this page's own URL had one (a token-gated box).
-        var parts = [];
-        if (token) parts.push("token=" + encodeURIComponent(token));
-        if (session) parts.push("session=" + encodeURIComponent(session));
-        Object.keys(o).forEach(function (k) { parts.push(k + "=" + encodeURIComponent(o[k])); });
-        // Author-declared knobs: knob.<key>=<value>. The server infers the type from the preview's
-        // declaration, so no <kind>: prefix. A knob still at its declared default is omitted — that
-        // keeps the URL on the instant baked snapshot (any knob.* param routes a published catalog
-        // to the daemon for a fresh re-render); only an actually-changed knob is sent.
-        document.querySelectorAll(".cp-knob").forEach(function (el) {
-          if (el.disabled) return;
-          var key = el.getAttribute("data-knob-key");
-          if (!key) return;
-          var val = (el.type === "checkbox") ? (el.checked ? "true" : "false") : el.value;
-          if (val === "") return;
-          if (val === (el.getAttribute("data-knob-initial") || "")) return;
-          parts.push("knob." + encodeURIComponent(key) + "=" + encodeURIComponent(val));
-        });
-        // Remote Compose knobs: rc.<name>=<kind>:<value>. The <kind>: prefix types the seed
-        // (color:%23AARRGGBB, int:…, bool:true, …). A knob still at its declared default is omitted
-        // so the URL stays on the instant baked snapshot until it's actually changed.
-        document.querySelectorAll(".cp-rc-knob").forEach(function (el) {
-          if (el.disabled) return;
-          var name = el.getAttribute("data-rc-name");
-          if (!name) return;
-          var kind = el.getAttribute("data-rc-kind") || "string";
-          var val = (el.type === "checkbox") ? (el.checked ? "true" : "false") : el.value;
-          if (val === "") return;
-          if (val === (el.getAttribute("data-rc-initial") || "")) return;
-          parts.push("rc." + encodeURIComponent(name) + "=" + encodeURIComponent(kind + ":" + val));
-        });
-        // App-declared theme (themeProvider = provider FQN). Routes to the daemon like a knob; a
-        // published catalog re-renders on demand. Omitted at "(default)" so the URL stays on the
-        // instant baked snapshot until a theme is actually chosen.
-        var tp = document.getElementById("cp-themeProvider");
-        if (tp && !tp.disabled && tp.value) {
-          parts.push("themeProvider=" + encodeURIComponent(tp.value));
-        }
-        // Detected-feature: keyboard focus (focus=0). Routes to the daemon like a knob; omitted when
-        // unchecked so the URL stays on the baked snapshot.
-        var fc = document.getElementById("cp-focus");
-        if (fc && !fc.disabled && fc.checked) parts.push("focus=0");
-        // Detected-feature: one-handed gesture hints (gestures=true). Routes to the daemon like a
-        // knob; omitted when unchecked so the URL stays on the baked snapshot.
-        var gc = document.getElementById("cp-gestures");
-        if (gc && !gc.disabled && gc.checked) parts.push("gestures=true");
-        // Remote Compose render backend: a server-side player pick rides the render as
-        // rcPlayer=<wire>. Emitted only once the visitor picks a backend (rcPlayerPicked) and only
-        // for a server-side lane — java / cmp-android render through the daemon, cmp-jvm through its
-        // isolated desktop subprocess (all three PNG lanes). The js canvas replays the doc in-browser
-        // (no server render), so it never sends the param, and an unpicked default stays on the
-        // instant baked snapshot.
-        if (
-          rcPlayerPicked &&
-          (rcPlayerBackend === "java" ||
-            rcPlayerBackend === "cmp-android" ||
-            rcPlayerBackend === "cmp-jvm")
-        ) {
-          parts.push("rcPlayer=" + encodeURIComponent(rcPlayerBackend));
-        }
-        return parts.join("&");
-      }
-      // "Full page (scroll)" appends `scroll=long` — but only to the SVG lane (the raster PNG has no
-      // full-page export), so the toggle scopes to `.svg` and leaves the PNG URL untouched.
-      var scrollLong = document.getElementById("cp-scroll-long");
-      function withScroll(ext, qs) {
-        if (ext === ".svg" && scrollLong && scrollLong.checked) {
-          return qs ? qs + "&scroll=long" : "scroll=long";
-        }
-        return qs;
-      }
-      function refreshSnapshot() {
-        status.textContent = "rendering…";
-        var qs = withScroll(snapshotExt, query());
-        var url =
-          base + "/render/" + encodeURIComponent(previewId) + snapshotExt + (qs ? "?" + qs : "");
-        var next = new Image();
-        next.onload = function () { img.src = url; status.textContent = ""; clearModeError(); };
-        next.onerror = function () {
-          showModeError((snapshotExt === ".svg" ? "SVG" : "PNG") + " render failed for this preview.");
-        };
-        next.src = url;
-        refreshLinks();
-      }
-      // The copyable direct-link panel: rebuild the absolute /render URLs (PNG + optional SVG) from
-      // the current controls so a copied/downloaded link reproduces exactly what's on screen. Built
-      // on location.origin so the link is absolute (curl-able / shareable), and kept in sync on
-      // every control or knob change — even the ones that don't re-render the snapshot themselves.
-      function renderUrl(ext) {
-        var qs = withScroll(ext, query());
-        return location.origin + base + "/render/" + encodeURIComponent(previewId) + ext +
-          (qs ? "?" + qs : "");
-      }
-      function withMode(url, mode) {
-        return url + (url.indexOf("?") >= 0 ? "&" : "?") + "mode=" + mode;
-      }
-      function refreshLinks() {
-        [["png", ".png"], ["svg", ".svg"]].forEach(function (pair) {
-          var field = document.getElementById("cp-url-" + pair[0]);
-          if (!field) return;
-          var embed = renderUrl(pair[1]);
-          var dl = document.getElementById("cp-dl-" + pair[0]);
-          if (pair[1] === ".svg") {
-            // Copy URL yields the web/document variant (`?mode=web` → external Google Fonts
-            // @import), so opening the copied link in a browser pulls the faces from Google. Copy
-            // SVG and the download stay on the embedded variant (self-contained — right for pasting
-            // into Figma or an <img>, where external refs don't load); the button reads it from
-            // data-embed-url.
-            field.value = withMode(embed, "web");
-            field.setAttribute("data-embed-url", embed);
-            if (dl) dl.href = embed;
-          } else {
-            field.value = embed;
-            if (dl) dl.href = embed;
-          }
-        });
-      }
-      // Click the read-only URL field to copy the /render URL to the clipboard (no separate button).
-      // The text is selected either way as a fallback + visible affordance, and the field flashes via
-      // .cp-url-copied so the click reads as "copied".
-      document.querySelectorAll(".cp-url").forEach(function (field) {
-        field.addEventListener("click", function () {
-          field.select();
-          var flash = function () {
-            field.classList.add("cp-url-copied");
-            setTimeout(function () { field.classList.remove("cp-url-copied"); }, 1200);
-          };
-          if (navigator.clipboard && navigator.clipboard.writeText) {
-            navigator.clipboard.writeText(field.value).then(flash, function () {});
-          } else {
-            try { document.execCommand("copy"); flash(); } catch (e) {}
-          }
-        });
-      });
-      // "Copy PNG" / "Copy SVG": fetch the current /render artefact and put it on the clipboard as
-      // text — SVG markup verbatim, PNG as a base64 data: URI — so it can be pasted straight into an
-      // editor, prompt, or issue without downloading a file. Uses the same live cp-url-<ext> field
-      // the URL Copy button reads, so the copied artefact matches the on-screen overrides.
-      document.querySelectorAll(".cp-copyimg").forEach(function (btn) {
-        btn.addEventListener("click", function () {
-          var field = document.getElementById(btn.getAttribute("data-copyimg-target"));
-          if (!field || !field.value) return;
-          var ext = btn.getAttribute("data-copyimg-ext");
-          var was = btn.getAttribute("data-copyimg-label") || btn.textContent;
-          btn.setAttribute("data-copyimg-label", was);
-          var reset = function (label) {
-            btn.textContent = label;
-            setTimeout(function () { btn.textContent = was; }, 1400);
-          };
-          if (!navigator.clipboard || !navigator.clipboard.writeText) { reset("No clipboard"); return; }
-          btn.textContent = "Copying…";
-          // Copy SVG targets the EMBEDDED variant (data-embed-url) — the field itself holds the
-          // web-mode URL for Copy URL, but a copied SVG is usually pasted into Figma / an editor,
-          // which needs the fonts baked in, not an external @import. PNG has one variant.
-          var src = (ext === ".svg" && field.getAttribute("data-embed-url")) || field.value;
-          // fetch() resolves even on a non-2xx render (503 saturated, 400 bad override, 404 a
-          // preview that can't export that lane), so guard on r.ok — otherwise the error body,
-          // not the artefact, would land on the clipboard and still report "Copied".
-          var okOrThrow = function (r) { if (!r.ok) throw new Error("render " + r.status); return r; };
-          var toText =
-            ext === ".svg"
-              ? fetch(src).then(okOrThrow).then(function (r) { return r.text(); })
-              : fetch(src)
-                  .then(okOrThrow)
-                  .then(function (r) { return r.blob(); })
-                  .then(function (blob) {
-                    return new Promise(function (resolve, reject) {
-                      var fr = new FileReader();
-                      fr.onload = function () { resolve(fr.result); };
-                      fr.onerror = function () { reject(fr.error); };
-                      fr.readAsDataURL(blob);
-                    });
-                  });
-          toText
-            .then(function (text) { return navigator.clipboard.writeText(text); })
-            .then(function () { reset("Copied"); }, function () { reset("Failed"); });
-        });
-      });
-      function drawFrame(b64, codec) {
-        var im = new Image();
-        im.onload = function () {
-          canvas.width = im.naturalWidth;
-          canvas.height = im.naturalHeight;
-          canvas.getContext("2d").drawImage(im, 0, 0);
-          // A <canvas> stretches its buffer to fill its CSS box, so a daemon frame whose aspect
-          // differs from the pinned snapshot box would squish. Cache the buffer dims and re-fit the
-          // element (contain, centred) so the frame letterboxes within the snapshot footprint
-          // instead of distorting to fill it.
-          liveW = im.naturalWidth;
-          liveH = im.naturalHeight;
-          fitLiveCanvas();
-        };
-        im.src = "data:image/" + (codec || "png") + ";base64," + b64;
-      }
-      // --- Live input forwarding (no-op on the snapshot lane). Coordinates are image-natural
-      // pixels; pointer events are grouped by pointerId so Compose's gesture pipeline tracks drags
-      // and multi-touch. Keys map to Android KEYCODE_* decimal strings (the daemon's wire format).
-      function liveActive() { return ws && ws.readyState === 1 && canvas.width; }
-      function sendInput(msg) {
-        if (liveActive()) ws.send(JSON.stringify(Object.assign({ type: "input" }, msg)));
-      }
-      function pixel(ev) {
-        var rect = canvas.getBoundingClientRect();
-        if (!rect.width || !rect.height) return null;
-        return {
-          x: Math.round((ev.clientX - rect.left) / rect.width * canvas.width),
-          y: Math.round((ev.clientY - rect.top) / rect.height * canvas.height),
-        };
-      }
-      // Per-pointer state. The pointerDown is *deferred* until the first move so a tap with no drag
-      // becomes a single `click` (matching the daemon's CLICK fast-path, which renders between press
-      // and release — a batched down+up can race Modifier.clickable). pointermove is coalesced to one
-      // send per pointerId per animation frame, so a fast drag doesn't flood the lane and concurrent
-      // fingers don't overwrite each other (multi-touch).
-      var pointers = {};           // pointerId -> { x, y, moved }
-      var pendingMoves = {};       // pointerId -> pointerMove message
-      var moveScheduled = false;
-      function flushMoves() {
-        moveScheduled = false;
-        var snapshot = pendingMoves;
-        pendingMoves = {};
-        Object.keys(snapshot).forEach(function (id) { sendInput(snapshot[id]); });
-      }
-      canvas.addEventListener("pointerdown", function (ev) {
-        if (!liveActive()) return;
-        var p = pixel(ev); if (!p) return;
-        canvas.focus();
-        try { canvas.setPointerCapture(ev.pointerId); } catch (e) {}
-        pointers[ev.pointerId] = { x: p.x, y: p.y, moved: false };
-      });
-      canvas.addEventListener("pointermove", function (ev) {
-        if (!liveActive() || ev.buttons === 0) return; // only while pressed (a drag)
-        var st = pointers[ev.pointerId]; if (!st) return;
-        var p = pixel(ev); if (!p) return;
-        if (!st.moved) {
-          // First movement → this is a drag: emit the deferred press at the original point.
-          st.moved = true;
-          sendInput({ kind: "pointerDown", pixelX: st.x, pixelY: st.y, pointerId: ev.pointerId });
-        }
-        pendingMoves[ev.pointerId] =
-          { kind: "pointerMove", pixelX: p.x, pixelY: p.y, pointerId: ev.pointerId };
-        if (!moveScheduled) { moveScheduled = true; requestAnimationFrame(flushMoves); }
-      });
-      function endPointer(ev) {
-        var st = pointers[ev.pointerId]; if (!st) return;
-        delete pointers[ev.pointerId];
-        var p = pixel(ev) || { x: st.x, y: st.y };
-        if (st.moved) {
-          flushMoves();
-          sendInput({ kind: "pointerUp", pixelX: p.x, pixelY: p.y, pointerId: ev.pointerId });
-        } else {
-          // No drag → a tap. Send a single CLICK (the daemon renders between press and release).
-          sendInput({ kind: "click", pixelX: st.x, pixelY: st.y, pointerId: ev.pointerId });
-        }
-      }
-      canvas.addEventListener("pointerup", endPointer);
-      canvas.addEventListener("pointercancel", endPointer);
-      canvas.addEventListener("wheel", function (ev) {
-        if (!liveActive()) return;
-        var p = pixel(ev); if (!p) return;
-        ev.preventDefault();
-        // Both daemon dispatchers drop non-key input without a position, so include the pixel.
-        sendInput({ kind: "rotaryScroll", pixelX: p.x, pixelY: p.y, scrollDeltaY: ev.deltaY });
-      }, { passive: false });
-      // Keyboard: focus the canvas (tabindex) to type. Maps the common keys to Android keycodes;
-      // unmapped keys are dropped (the daemon ignores codes outside its translation table anyway).
-      canvas.tabIndex = 0;
-      function androidKeycode(k) {
-        if (k.length === 1) {
-          var c = k.toLowerCase();
-          if (c >= "a" && c <= "z") return String(29 + (c.charCodeAt(0) - 97)); // KEYCODE_A = 29
-          if (c >= "0" && c <= "9") return String(7 + (c.charCodeAt(0) - 48)); // KEYCODE_0 = 7
-          if (k === " ") return "62"; // SPACE
-        }
-        switch (k) {
-          case "Enter": return "66";
-          case "Backspace": return "67";
-          case "Tab": return "61";
-          case "Escape": return "111";
-          case "Delete": return "112";
-          case "ArrowUp": return "19";
-          case "ArrowDown": return "20";
-          case "ArrowLeft": return "21";
-          case "ArrowRight": return "22";
-          default: return null;
-        }
-      }
-      function keyInput(kind, ev) {
-        if (!liveActive()) return;
-        var code = androidKeycode(ev.key);
-        if (code === null) return;
-        ev.preventDefault();
-        sendInput({ kind: kind, keyCode: code });
-      }
-      canvas.addEventListener("keydown", function (ev) { keyInput("keyDown", ev); });
-      canvas.addEventListener("keyup", function (ev) { keyInput("keyUp", ev); });
-      function openStream() {
-        root.setAttribute("data-mode", "live");
-        // Seed the canvas buffer with the current snapshot *before* the swap, so there's no blank
-        // flash while "connecting…" (the first frame overwrites it).
-        if (img.naturalWidth && img.naturalHeight) {
-          canvas.width = img.naturalWidth;
-          canvas.height = img.naturalHeight;
-          try { canvas.getContext("2d").drawImage(img, 0, 0); } catch (e) {}
-        }
-        // Mount the canvas as an absolute overlay on the snapshot's slot — the same fixed box the
-        // Wasm tier locks to. The img stays in flow (visibility:hidden keeps its slot), so the stage
-        // geometry is defined once by the snapshot and a live frame whose pixel dims differ from the
-        // baked PNG scales into this box instead of resizing the stage (drawFrame only touches the
-        // buffer now, never the layout). Input mapping reads the buffer size, so it's unaffected.
-        canvas.classList.add("cp-canvas-live");
-        positionOverlay(canvas);
-        img.style.visibility = "hidden";
-        canvas.hidden = false;
-        setPending("connecting…");
-        var proto = location.protocol === "https:" ? "wss:" : "ws:";
-        // Request WebP frames (smaller; the browser decodes them via the data URL, and the daemon
-        // downgrades to PNG when it can't encode WebP — each frame carries its actual codec).
-        var qs = query();
-        // Track whether the stream ever delivered a frame: a close/error *before* the first frame is
-        // a failed activation (surface it), whereas a close *after* frames is just a normal teardown.
-        var liveGotFrame = false;
-        // Hold the socket in a per-activation local as well as `ws`, and gate every callback on
-        // `ws === sock`. Toggling Live off and straight back on opens a replacement before the old
-        // socket's close event is delivered, and that stale callback would otherwise clear the NEW
-        // connection's pending badge (its own liveGotFrame is true, so it skips the error branch)
-        // and null out `ws` — orphaning the live socket the input/override senders reach for.
-        var sock = new WebSocket(proto + "//" + location.host + base + "/ws/" +
-          encodeURIComponent(previewId) + "?" + (qs ? qs + "&codec=webp" : "codec=webp"));
-        ws = sock;
-        sock.onopen = function () {
-          // The connect URL seeds only query()'s fields — the display axes plus changed knobs — so
-          // the live-only overlays (talkBack / touchOverlay) and anything toggled during the
-          // connecting window aren't in it. Replay the full live override map once the socket is
-          // ready so the daemon reflects the exact current control state, including an overlay
-          // checked before onopen whose change event the readyState guard dropped.
-          sock.send(JSON.stringify({ type: "setOverrides", overrides: liveOverrides() }));
-        };
-        sock.onmessage = function (ev) {
-          // A frame from a socket the viewer has already replaced is stale in both senses: it must
-          // not paint over the new lane's stage, nor report it as connected.
-          if (ws !== sock) return;
-          var m;
-          try { m = JSON.parse(ev.data); } catch (e) { return; }
-          if (m.type === "frame") { liveGotFrame = true; clearModeError(); drawFrame(m.dataBase64, m.codec); setPending(null); }
-          else if (m.type === "error") { showModeError(m.message || "Live preview error."); }
-        };
-        // onerror always precedes onclose; let onclose decide (it carries the code/reason). Only
-        // surface here if the socket somehow errors while already open+frame-less and never closes.
-        sock.onerror = function () { if (ws === sock && !liveGotFrame) setPending("connecting…"); };
-        sock.onclose = function (ev) {
-          // Not the current socket ⇒ a teardown the viewer already accounted for in closeStream()
-          // (which cleared pending and restored the snapshot). Leave the live lane's state alone.
-          if (ws !== sock) return;
-          ws = null;
-          // The lane is done waiting either way — it painted, or it failed (showModeError below).
-          setPending(null);
-          // Closed before any frame ⇒ the mode failed to activate. Drop the stale seeded snapshot
-          // from the canvas so it can't masquerade as a live render, and surface why.
-          if (!liveGotFrame && live && live.checked) {
-            canvas.hidden = true;
-            img.style.removeProperty("visibility");
-            showModeError(liveCloseReason(ev));
-          }
-        };
-      }
-      function closeStream() {
-        root.setAttribute("data-mode", "snapshot");
-        // Toggling Live off mid-connect must not leave the badge stuck on "connecting…".
-        setPending(null);
-        if (ws) { ws.close(); ws = null; }
-        canvas.hidden = true;
-        // Tear down the overlay: drop the absolute positioning and restore the snapshot img's slot.
-        canvas.classList.remove("cp-canvas-live");
-        canvas.style.removeProperty("left"); canvas.style.removeProperty("top");
-        canvas.style.removeProperty("width"); canvas.style.removeProperty("height");
-        // Forget the last frame's dims so a reconnect seeds from the snapshot (fill) until its own
-        // first frame re-fits, rather than briefly letterboxing to a stale aspect.
-        liveW = 0;
-        liveH = 0;
-        img.style.removeProperty("visibility");
-        img.hidden = false;
-      }
-      // --- Wasm tier (the in-browser CMP app, mounted in a sandboxed iframe). Only wired when the
-      // session carries a Wasm app (data-wasm-src present). Theme/font-scale/locale re-point the
-      // iframe's ?uiMode/?fontScale/?localeTag (device/orientation are server-render-only).
-      var stage = document.querySelector(".cp-stage");
-      var wasmFrame = document.getElementById("cp-wasm");
-      var wasmToggle = document.getElementById("cp-wasm-toggle");
-      var wasmSrc = root.getAttribute("data-wasm-src") || "";
-      // Set once the app has painted its first frame (its "cp-wasm-ready" message). Until then a
-      // control change re-points ?query (initial load); after, it posts an override patch so the
-      // app recomposes in place instead of reloading the whole ~20 MB Wasm bundle.
-      var wasmReady = false;
-      // Boot watchdog: if the app never signals "cp-wasm-ready" (bundle 404, Wasm/GL failure, …),
-      // surface a visible error instead of leaving the stage stuck on "loading Wasm…" forever.
-      var wasmBootTimer = null;
-      function wasmBaseSrc() {
-        if (!wasmSrc) return "";
-        // The src comes from a server-set data- attribute, but resolve it against our own origin and
-        // refuse anything not same-origin http(s) anyway — so a `javascript:`/`data:` URL can never
-        // reach the iframe even if the attribute were ever mis-set (defuses DOM-text-as-HTML). The
-        // query is left as the server baked it (the variant's default theme) — session overrides
-        // never go in it, so it stays the app's clean base to revert to when a control is cleared.
-        var u;
-        try { u = new URL(wasmSrc, location.origin); } catch (e) { return ""; }
-        if (u.origin !== location.origin) return "";
-        return u.href;
-      }
-      // NOTE: the font prefetch lives in the app's own index.html (it starts the manifest+font
-      // fetches at document load, in parallel with the Wasm boot), not on this page. That's where
-      // it belongs regardless of the sandbox: it must be in flight before the iframe navigates, and
-      // the app is the one that consumes the promises. (Historically the iframe was opaque-origin
-      // with its own cache partition, so a page-side preload was also unreusable and fetched every
-      // font twice; with allow-same-origin the partition is shared, but the app-side prefetch is
-      // still the right home, so keep page-side preloads out — see the ServeWebFixtureTest guard.)
-      // The override patch (theme / font scale / locale) the running app merges over its baked base —
-      // a bare `a=b&c=d` query. An absent key falls back to the app's baked default (e.g. cleared
-      // Theme → the variant's uiMode). Device / orientation are server-render-only, so not forwarded.
-      // The stage checkerboard's tile origin in the iframe's own CSS-px coordinates. The app can't
-      // render a transparent surface, so it paints this same pattern itself; handing it the phase
-      // makes the in-canvas cells continue the page's cells exactly (the stylesheet positions the
-      // 16px tile at 50% of the stage's padding box; the iframe sits at style.left/top within it).
-      function wasmBgPhase() {
-        var left = parseFloat(wasmFrame.style.left) || 0;
-        var top = parseFloat(wasmFrame.style.top) || 0;
-        var x = (stage.clientWidth - 16) / 2 - left;
-        var y = (stage.clientHeight - 16) / 2 - top;
-        return x.toFixed(2) + "," + y.toFixed(2);
-      }
-      function wasmOverridePatch() {
-        var parts = [];
-        var el = document.getElementById("cp-uiMode");
-        if (el && el.value) parts.push("uiMode=" + encodeURIComponent(el.value));
-        var loc = document.getElementById("cp-localeTag");
-        if (loc && loc.value) parts.push("localeTag=" + encodeURIComponent(loc.value));
-        if (fontScaleTouched && fs) parts.push("fontScale=" + encodeURIComponent(fs.value));
-        parts.push("bgPhase=" + encodeURIComponent(wasmBgPhase()));
-        // Author-declared knobs also apply in the browser: the wasm catalog seeds its
-        // `catalogOverride*` from these `knob.<key>` params. Mirror query() — omit a knob still at
-        // its `data-knob-initial` so an unedited knob reverts to the author default in the app.
-        document.querySelectorAll(".cp-knob").forEach(function (el) {
-          if (el.disabled) return;
-          var key = el.getAttribute("data-knob-key");
-          if (!key) return;
-          var val = (el.type === "checkbox") ? (el.checked ? "true" : "false") : el.value;
-          if (val === "") return;
-          if (val === (el.getAttribute("data-knob-initial") || "")) return;
-          parts.push("knob." + encodeURIComponent(key) + "=" + encodeURIComponent(val));
-        });
-        return parts.join("&");
-      }
-      // Initial iframe URL: the baked base plus the current overrides in the `#…` fragment, so the
-      // app's first paint honours them yet keeps the query as its true base (a later clear reverts).
-      function wasmInitialSrc() {
-        var base = wasmBaseSrc();
-        if (!base) return "";
-        var patch = wasmOverridePatch();
-        return patch ? base + "#" + patch : base;
-      }
-      // Pixel parity: lay an absolute overlay ([el] — the Wasm iframe or the live canvas) exactly
-      // over the snapshot's rendered box, so switching to it shouldn't move anything. Both the Wasm
-      // app (contain-fitting the same sticker geometry the snapshot baked) and the daemon (the same
-      // preview re-rendered) fill this box, so the three transports share one geometry.
-      function positionOverlay(el) {
-        var sr = stage.getBoundingClientRect();
-        var r = img.getBoundingClientRect();
-        if (r.width > 0 && r.height > 0) {
-          // Offsets are relative to the stage's padding box — subtract its border (clientLeft/Top).
-          el.style.left = (r.left - sr.left - stage.clientLeft) + "px";
-          el.style.top = (r.top - sr.top - stage.clientTop) + "px";
-          el.style.width = r.width + "px";
-          el.style.height = r.height + "px";
-        } else {
-          // No snapshot box to mirror (e.g. its render 404'd): fill the stage's content box.
-          el.style.left = "12px"; el.style.top = "12px";
-          el.style.width = "calc(100% - 24px)";
-          el.style.height = (stage.clientHeight - 24) + "px";
-        }
-      }
-      function positionWasmFrame() { positionOverlay(wasmFrame); }
-      // The live canvas can't just fill the snapshot box like the Wasm frame does: a <canvas>
-      // stretches its buffer to its CSS box, so pinning a differently-shaped daemon frame to the
-      // snapshot's rect squished the render. Fit the frame (contain) inside that rect, centred — it
-      // letterboxes within the snapshot footprint (so the stage still never resizes, the property
-      // the pinned box was introduced for) instead of distorting. liveW/liveH cache the current
-      // buffer so a window resize re-fits; unset (before the first frame, when the buffer is seeded
-      // from the same-aspect snapshot) it fills the box exactly, matching positionOverlay.
-      var liveW = 0;
-      var liveH = 0;
-      function fitLiveCanvas() {
-        var sr = stage.getBoundingClientRect();
-        var r = img.getBoundingClientRect();
-        var boxLeft, boxTop, boxW, boxH;
-        if (r.width > 0 && r.height > 0) {
-          boxLeft = r.left - sr.left - stage.clientLeft;
-          boxTop = r.top - sr.top - stage.clientTop;
-          boxW = r.width;
-          boxH = r.height;
-        } else {
-          boxLeft = 12;
-          boxTop = 12;
-          boxW = stage.clientWidth - 24;
-          boxH = stage.clientHeight - 24;
-        }
-        var w = boxW;
-        var h = boxH;
-        if (liveW > 0 && liveH > 0) {
-          var scale = Math.min(boxW / liveW, boxH / liveH);
-          w = liveW * scale;
-          h = liveH * scale;
-        }
-        canvas.style.left = (boxLeft + (boxW - w) / 2) + "px";
-        canvas.style.top = (boxTop + (boxH - h) / 2) + "px";
-        canvas.style.width = w + "px";
-        canvas.style.height = h + "px";
-      }
-      // Swap the stage from the snapshot to the (already-painted) Wasm frame. The snapshot keeps
-      // its layout slot (visibility, not display) so the stage geometry — and the overlay tracking
-      // it — never shifts.
-      function revealWasm() {
-        if (!wasmActive() || wasmReady) return;
-        wasmReady = true;
-        if (wasmBootTimer) { clearTimeout(wasmBootTimer); wasmBootTimer = null; }
-        clearModeError();
-        positionWasmFrame();
-        wasmFrame.classList.add("cp-wasm-live");
-        img.style.visibility = "hidden";
-        status.textContent = "";
-        // Re-sync any control changed during load (the fragment only captured open-time state).
-        var patch = wasmOverridePatch();
-        if (patch && wasmFrame.contentWindow) wasmFrame.contentWindow.postMessage(patch, "*");
-      }
-      function openWasm() {
-        // No-op without a Wasm app (wasmFrame absent): enterMode() calls this unconditionally, but a
-        // non-Wasm daemon/static session has no iframe to drive. Guard so touching it can't throw.
-        if (!wasmFrame) return;
-        // Wasm and the daemon stream are mutually exclusive; the mode switch (enterMode) tears the
-        // stream down before opening Wasm, so there's nothing extra to close here.
-        root.setAttribute("data-mode", "wasm");
-        canvas.hidden = true;
-        // Keep the snapshot visible while the app loads; the iframe mounts over it at opacity 0
-        // and only fades in on the app's first-frame signal — no blank/white flash.
-        positionWasmFrame();
-        wasmFrame.hidden = false;
-        wasmReady = false;
-        wasmFrame.src = wasmInitialSrc();
-        status.textContent = "loading Wasm…";
-        if (wasmBootTimer) clearTimeout(wasmBootTimer);
-        wasmBootTimer = setTimeout(function () {
-          if (!wasmReady && wasmActive()) {
-            showModeError("Wasm preview didn't start — the in-browser app failed to load.");
-          }
-        }, 20000);
-      }
-      function closeWasm() {
-        // No Wasm iframe (non-Wasm session) ⇒ nothing to tear down; enterMode() still calls this
-        // unconditionally when switching to Live/PNG, so guard against the null frame.
-        if (!wasmFrame) return;
-        root.setAttribute("data-mode", "snapshot");
-        wasmReady = false;
-        if (wasmBootTimer) { clearTimeout(wasmBootTimer); wasmBootTimer = null; }
-        wasmFrame.classList.remove("cp-wasm-live");
-        wasmFrame.hidden = true; wasmFrame.removeAttribute("src");
-        img.style.removeProperty("visibility");
-        img.hidden = false;
-        status.textContent = "";
-      }
-      function wasmActive() { return wasmToggle && wasmToggle.checked; }
-
-      // ---- In-browser Remote Compose canvas lane ----------------------------------------------
-      // When this preview carries a captured `.rc` document, the "RC (browser)" toggle paints it with
-      // the vendored player (RC.RcdPlayer) into #cp-rc-canvas — no daemon — and Remote Compose knob
-      // edits apply live via setNamed*Override + repaint (onRcKnobChanged) instead of a /render
-      // round-trip. Opt-in like Live / Wasm, so the default PNG snapshot is untouched.
-      var rcCanvasEl = document.getElementById("cp-rc-canvas");
-      var rcToggle = document.getElementById("cp-rc-toggle");
-      var hasRcDoc = root.getAttribute("data-has-rc-doc") === "1";
-      var rcBtn = document.getElementById("cp-rc-btn");
-      var rcPlayer = null; // the RC.RcdPlayer instance (created lazily on first open)
-      var rcCtx = null; // its WebRemoteContext, for named-value overrides
-      var rcReady = false; // a first frame is painted and the canvas revealed
-      var rcScriptState = 0; // 0 = not loaded, 1 = loading, 2 = ready
-      var rcScriptWaiters = [];
-      function rcAvailable() { return !!(hasRcDoc && rcCanvasEl); }
-      function rcActive() { return !!(rcToggle && rcToggle.checked); }
-      // Lazy-load the shared player bundle once (a constant, session-independent path); queue callers
-      // while it loads so a fast re-open can't inject the script twice.
-      function ensureRcScript(cb) {
-        if (rcScriptState === 2 || window.RC) { rcScriptState = 2; cb(true); return; }
-        rcScriptWaiters.push(cb);
-        if (rcScriptState === 1) return;
-        rcScriptState = 1;
-        var s = document.createElement("script");
-        s.src = "/rc-player/bundle.js";
-        s.onload = function () {
-          rcScriptState = 2;
-          var ws = rcScriptWaiters; rcScriptWaiters = [];
-          ws.forEach(function (f) { f(true); });
-        };
-        s.onerror = function () {
-          rcScriptState = 0;
-          var ws = rcScriptWaiters; rcScriptWaiters = [];
-          ws.forEach(function (f) { f(false); });
-        };
-        document.head.appendChild(s);
-      }
-      // The `.rc` document URL — the same `base` + token/session as the snapshot, but no override qs
-      // (the lane serves the document verbatim; knob edits apply client-side).
-      function rcDocUrl() {
-        var parts = [];
-        if (token) parts.push("token=" + encodeURIComponent(token));
-        if (session) parts.push("session=" + encodeURIComponent(session));
-        var qs = parts.join("&");
-        return base + "/render/" + encodeURIComponent(previewId) + ".rc" + (qs ? "?" + qs : "");
-      }
-      // Parse #RRGGBB / #AARRGGBB (optionally %23-escaped) into a 0xAARRGGBB int (opaque when no alpha).
-      function parseRcColor(v) {
-        if (!v) return null;
-        var h = v.replace(/^%23/, "").replace(/^#/, "");
-        if (h.length === 6) h = "FF" + h;
-        if (h.length !== 8) return null;
-        var n = parseInt(h, 16);
-        return isNaN(n) ? null : (n >>> 0);
-      }
-      // Push every Remote Compose knob's current value onto the player's context, then repaint. Names
-      // are USER:-domain-qualified (the connector registers author knobs under USER:), matching the
-      // document's named variables; kinds mirror query()'s rc.<name>=<kind>:<value> typing.
-      function applyRcOverrides() {
-        if (!rcCtx) return;
-        document.querySelectorAll(".cp-rc-knob").forEach(function (el) {
-          var name = el.getAttribute("data-rc-name");
-          if (!name) return;
-          var qn = "USER:" + name;
-          var kind = el.getAttribute("data-rc-kind") || "string";
-          var val = (el.type === "checkbox") ? (el.checked ? "true" : "false") : el.value;
-          try {
-            if (kind === "color") {
-              var argb = parseRcColor(val);
-              if (argb !== null && rcCtx.setNamedColorOverride) rcCtx.setNamedColorOverride(qn, argb);
-            } else if (kind === "float" || kind === "dp") {
-              var f = parseFloat(val);
-              if (!isNaN(f) && rcCtx.setNamedFloatOverride) rcCtx.setNamedFloatOverride(qn, f);
-            } else if (kind === "int" || kind === "integer") {
-              var n = parseInt(val, 10);
-              if (!isNaN(n) && rcCtx.setNamedIntegerOverride) rcCtx.setNamedIntegerOverride(qn, n);
-            } else if (kind === "bool" || kind === "boolean") {
-              // The player's setNamedBooleanOverride only records the value — it doesn't touch the
-              // render state — so route booleans through the integer setter as 1/0, matching the
-              // daemon's BooleanValue → user-local-integer mapping.
-              if (rcCtx.setNamedIntegerOverride) {
-                rcCtx.setNamedIntegerOverride(qn, val === "true" ? 1 : 0);
-              }
-            } else if (rcCtx.setNamedStringOverride) {
-              rcCtx.setNamedStringOverride(qn, val);
-            }
-          } catch (e) { /* a knob the document doesn't declare is a harmless no-op */ }
-        });
-        if (rcPlayer && rcPlayer.repaint) rcPlayer.repaint();
-      }
-      function openRc() {
-        if (!rcAvailable()) return;
-        root.setAttribute("data-mode", "rc");
-        canvas.hidden = true;
-        rcReady = false;
-        status.textContent = "loading RC player…";
-        ensureRcScript(function (ok) {
-          if (!ok || !window.RC) { showModeError("The Remote Compose player failed to load."); return; }
-          if (!rcActive()) return; // toggled away while the script loaded
-          fetch(rcDocUrl())
-            .then(function (r) { if (!r.ok) throw new Error("doc " + r.status); return r.arrayBuffer(); })
-            .then(function (buf) {
-              if (!rcActive()) return null;
-              // Size the canvas to the preview's real pixel dimensions BEFORE loading: the player
-              // derives the document viewport from the canvas's current size at load time, and a
-              // resize afterwards can't recover it. The baked snapshot <img> carries those
-              // dimensions (rendered at the same density), so a non-default-shaped preview fills
-              // the canvas instead of being letterboxed into the 300×150 default.
-              var w = img.naturalWidth || 0, h = img.naturalHeight || 0;
-              if (w > 0 && h > 0) { rcCanvasEl.width = w; rcCanvasEl.height = h; }
-              if (!rcPlayer) rcPlayer = new window.RC.RcdPlayer(rcCanvasEl);
-              return rcPlayer.loadFromArrayBuffer(buf);
-            })
-            .then(function () {
-              if (!rcActive()) return;
-              rcCtx = rcPlayer.getRemoteContext ? rcPlayer.getRemoteContext() : null;
-              applyRcOverrides();
-              if (rcPlayer.repaint) rcPlayer.repaint();
-              revealRc();
-            })
-            .catch(function () { showModeError("Rendering the Remote Compose document failed."); });
-        });
-      }
-      // Swap the stage from the snapshot to the painted canvas. The snapshot is removed from flow
-      // (display:none) so the stage takes the document's own size rather than stacking both.
-      function revealRc() {
-        if (!rcActive() || rcReady) return;
-        rcReady = true;
-        clearModeError();
-        rcCanvasEl.hidden = false;
-        img.style.display = "none";
-        if (rcBtn) rcBtn.setAttribute("aria-pressed", "true");
-        status.textContent = "";
-      }
-      function closeRc() {
-        if (!rcCanvasEl) return;
-        root.setAttribute("data-mode", "snapshot");
-        rcReady = false;
-        rcCanvasEl.hidden = true;
-        img.style.removeProperty("display");
-        img.hidden = false;
-        if (rcBtn) rcBtn.setAttribute("aria-pressed", "false");
-      }
-
-      function onControlsChanged() {
-        // Keep the copyable direct links current no matter which transport handles the change.
-        refreshLinks();
-        if (wasmActive()) {
-          // Recompose in place once the app is up; before it's ready, re-point the initial src (the
-          // fragment carries the overrides) — the load handler re-syncs the final state either way.
-          if (wasmReady && wasmFrame.contentWindow) {
-            wasmFrame.contentWindow.postMessage(wasmOverridePatch(), "*");
-          } else {
-            wasmFrame.src = wasmInitialSrc();
-          }
-          return;
-        }
-        if (live.checked && ws && ws.readyState === 1) {
-          ws.send(JSON.stringify({ type: "setOverrides", overrides: liveOverrides() }));
-          return;
-        }
-        // Not in an interactive lane. Whenever the server can produce a fresh overridden render — a
-        // live daemon session (!staticSnapshot) OR a published catalog whose carried daemon
-        // re-renders on demand (canRenderOverrides) — just re-point /render. This is what lets Size,
-        // Locale, Device, … take effect for a CMP catalog while still showing static snapshots, so
-        // the controls aren't dead until a live stream is opened.
-        if (!staticSnapshot || canRenderOverrides) { refreshSnapshot(); return; }
-        // Pure static published catalog whose only interactive lane is the in-browser app: the
-        // wasm-honoured controls (day/night, font scale, locale) can only apply in the browser, so
-        // auto-enable the Wasm tier and let it apply the change, instead of a dead /render the
-        // catalog can't serve. (Size/Device stay disabled here — the Wasm app can't honour them.)
-        if (wasmToggle) { setMode("wasm"); return; }
-        refreshSnapshot();
-      }
-
-      // The single Static⇄Live toggle drives these transports. "live" opens the daemon stream,
-      // "wasm" mounts the in-browser app, "png" (the default) is the static snapshot. closeStream /
-      // closeWasm are idempotent, so a switch safely tears down both regardless of the prior state.
-      // The static lane additionally honours the SVG format toggle: the same <img>, pointed at the
-      // vector `/render/<id>.svg` instead of the raster `.png`. A live lane (stream / wasm) is raster
-      // frames, so entering it clears SVG.
-      var svgToggle = document.getElementById("cp-svg-toggle");
-      function svgOn() {
-        return !!(svgToggle && svgToggle.getAttribute("aria-pressed") === "true");
-      }
-      function enterMode(m) {
-        // A mode switch always clears a prior lane's error; the new lane re-raises its own if it fails.
-        clearModeError();
-        if (m === "live") {
-          snapshotExt = ".png";
-          if (svgToggle) svgToggle.setAttribute("aria-pressed", "false");
-          closeWasm();
-          closeRc();
-          openStream();
-        } else if (m === "wasm") {
-          snapshotExt = ".png";
-          if (svgToggle) svgToggle.setAttribute("aria-pressed", "false");
-          closeStream();
-          closeRc();
-          openWasm();
-        } else if (m === "rc") {
-          snapshotExt = ".png";
-          if (svgToggle) svgToggle.setAttribute("aria-pressed", "false");
-          closeStream();
-          closeWasm();
-          openRc();
-        } else {
-          closeStream();
-          closeWasm();
-          closeRc();
-          // Static snapshot lane: raster PNG, or the vector SVG when the format toggle is on.
-          snapshotExt = svgOn() ? ".svg" : ".png";
-          if (svgOn()) root.setAttribute("data-mode", "svg");
-        }
-        syncOverlayToggles();
-        syncServerControls();
-        updateLiveToggle();
-        // Every lane transition passes through here, so the backend chips are re-reconciled whether
-        // the viewer entered/left the JS canvas via a chip or via Live/Wasm/SVG/snapshot controls.
-        syncRcBackendChips();
-        // Render the static lane AFTER syncServerControls() has reconciled the daemon-only controls
-        // for the new lane. Returning from Wasm this re-enables the `.cp-rc-knob` inputs first, so
-        // query() includes an rc.* value edited before the Wasm detour in the first snapshot render
-        // (and its direct links) instead of skipping the still-disabled control. The live/wasm lanes
-        // drive their own render (openStream / openWasm), so only the static lane renders here.
-        if (m !== "live" && m !== "wasm" && m !== "rc") refreshSnapshot();
-      }
-      // SVG format toggle: swap the static snapshot between raster and vector. Pressing it while a
-      // live lane is active drops back to the static vector render; pressing it in the static lane
-      // swaps the extension in place.
-      if (svgToggle) {
-        svgToggle.addEventListener("click", function () {
-          var turnOn = !svgOn();
-          svgToggle.setAttribute("aria-pressed", turnOn ? "true" : "false");
-          if (turnOn && (live.checked || wasmActive())) {
-            setMode("png"); // enterMode("png") reads svgOn() → renders the .svg
-          } else {
-            snapshotExt = turnOn ? ".svg" : ".png";
-            root.setAttribute("data-mode", turnOn ? "svg" : "snapshot");
-            refreshSnapshot();
-          }
-        });
-      }
-      // "Full page (scroll)" re-renders the on-screen SVG when it's the active format, and always
-      // reshapes the copyable/downloadable SVG export URL (withScroll scopes it to the `.svg` lane).
-      if (scrollLong) {
-        scrollLong.addEventListener("change", function () {
-          if (snapshotExt === ".svg") refreshSnapshot();
-          else refreshLinks();
-        });
-      }
-      // The live-only overlay toggles (talkBack / touchOverlay) are meaningful only while the daemon
-      // holds the composition, so they're enabled iff the live stream is the active mode and greyed
-      // out otherwise. Called on every mode transition.
-      var overlayToggles = document.querySelectorAll(".cp-overlay");
-      function syncOverlayToggles() {
-        var on = !!(live && live.checked);
-        Array.prototype.forEach.call(overlayToggles, function (el) { el.disabled = !on; });
-      }
-      // Enable/disable the display controls to match what the active session can actually render.
-      // A server-render control (Size / Device / Orientation / Background) takes effect whenever the
-      // server can produce a fresh overridden render: a live daemon session (!staticSnapshot), a
-      // catalog whose carried daemon re-renders on demand (canRenderOverrides), or an active live
-      // stream. The wasm-honoured trio (Day/Night / Locale / Font scale) additionally applies in the
-      // in-browser app, so it's also enabled whenever a Wasm app backs the session. This is what
-      // makes "most override modes" live for a CMP catalog (compose-m3) instead of greyed out until
-      // a stream is opened; the server-rendered markup already reflects this, and this keeps it in
-      // sync across mode transitions.
-      var serverOnlyControlIds =
-        ["device", "orientation", "background", "sizeMode",
-         "fixedW", "fixedH", "minW", "minH", "maxW", "maxH"];
-      var wasmHonouredControlIds = ["uiMode", "localeTag", "fontScale"];
-      var alwaysDark = root.getAttribute("data-always-dark") === "1";
-      function syncServerControls() {
-        // The in-browser Wasm lane only honours the wasm-honoured trio (uiMode/locale/fontScale) +
-        // knobs (see wasmOverridePatch); size/device/orientation/background and the app-theme
-        // selector re-point /render, which the iframe ignores. So while Wasm is the active lane they
-        // are dead — disable them (even on a catalog that can otherwise re-render) and restore them
-        // when the lane leaves Wasm. Called on every mode transition, so the states track the lane.
-        var onWasm = wasmActive();
-        // The RC canvas lane, like Wasm, honours only its own overrides (the Remote Compose knobs,
-        // applied client-side): size/device/locale/theme all re-point /render, which the painted
-        // canvas ignores. So it's as "dead" for the server + wasm-honoured controls as the Wasm lane.
-        var onRc = rcActive();
-        var canServerRender =
-          !onWasm && !onRc && (!staticSnapshot || canRenderOverrides || !!(live && live.checked));
-        serverOnlyControlIds.forEach(function (id) {
-          var el = document.getElementById("cp-" + id);
-          if (el) el.disabled = !canServerRender;
-        });
-        // The wasm-honoured trio stays live in the in-browser Wasm lane (the app applies it), so it's
-        // enabled whenever the server can render OR a Wasm app backs the session — but not in the RC
-        // canvas lane, which doesn't map them onto the document.
-        wasmHonouredControlIds.forEach(function (id) {
-          var el = document.getElementById("cp-" + id);
-          if (el) el.disabled = (id === "uiMode" && alwaysDark) ||
-            !(canServerRender || (wasmSrc && !onRc));
-        });
-        // The app-theme (themeProvider) selector is a daemon-only override the Wasm app doesn't
-        // honour: dead in the in-browser lane. Disable it there; otherwise mirror its server-side
-        // gate (enabled iff the host can render an override — canApplyOverrides || canRenderOverrides,
-        // i.e. !staticSnapshot || canRenderOverrides).
-        var themeProviderEl = document.getElementById("cp-themeProvider");
-        if (themeProviderEl) {
-          themeProviderEl.disabled = onWasm || onRc || !(!staticSnapshot || canRenderOverrides);
-        }
-        // Remote Compose knobs are LIVE in the RC canvas lane — an edit applies client-side via
-        // setNamed*Override + repaint (onRcKnobChanged), no daemon needed — so enable them whenever
-        // that lane is active. Outside it they're daemon-only (Remote Compose has no in-browser
-        // runtime), so, like the app-theme selector, they're dead in the Wasm lane and otherwise
-        // gated on the host being able to render an override.
-        document.querySelectorAll(".cp-rc-knob").forEach(function (el) {
-          el.disabled = onRc ? false : (onWasm || !(!staticSnapshot || canRenderOverrides));
-        });
-      }
-      // Programmatic switch (the live toggle, or a wasm-only control auto-enabling Wasm): tick the
-      // hidden mode radio so its state is consistent, then run the transition.
-      function setMode(m) {
-        var r = document.getElementById(
-          m === "live" ? "cp-live" :
-          m === "wasm" ? "cp-wasm-toggle" :
-          m === "rc" ? "cp-rc-toggle" : "cp-mode-png");
-        if (r) r.checked = true;
-        enterMode(m);
-      }
-      Array.prototype.forEach.call(
-        document.querySelectorAll("input[name=\"cp-mode\"]"),
-        function (r) {
-          r.addEventListener("change", function () { if (r.checked) enterMode(r.value); });
-        });
-      // The single Static⇄Live preview toggle. Off = the baked / on-demand snapshot; on = the best
-      // interactive lane this session offers (the daemon stream when present, else the in-browser
-      // Wasm app). Overrides still take effect while static (a catalog re-renders /render on
-      // demand), so this toggle is specifically about *interacting* with the running composition —
-      // clicking, scrolling, typing. The corner backend badge flips its icon/accent to match (see
-      // backendBadgeScript).
-      var liveToggle = document.getElementById("cp-live-toggle");
-      var liveToggleLabel = document.getElementById("cp-live-toggle-label");
-      // The dedicated in-browser Wasm toggle (present only when the session has BOTH a daemon lane
-      // and a Wasm app). When it's here, "Live preview" owns the daemon lane and this owns the Wasm
-      // lane, so the Wasm tier is reachable rather than hidden behind bestLiveMode()'s daemon
-      // preference. Null in the daemon-only / wasm-only cases, where the single toggle stays generic.
-      var wasmBtn = document.getElementById("cp-wasm-btn");
-      var modeHint = document.getElementById("cp-mode-hint");
-      function liveTransportAvailable() { return (live && !live.disabled) || !!wasmToggle; }
-      function bestLiveMode() { return (live && !live.disabled) ? "live" : (wasmToggle ? "wasm" : null); }
-      function anyLiveActive() { return !!(live && live.checked) || !!(wasmToggle && wasmToggle.checked); }
-      function updateLiveToggle() {
-        var liveOn = !!(live && live.checked);
-        var wasmOn = !!(wasmToggle && wasmToggle.checked);
-        if (liveToggle) {
-          // With a separate Wasm button, "Live preview" reflects the daemon lane alone; without one
-          // it's the generic interactive toggle that lights for either lane.
-          var livePressed = wasmBtn ? liveOn : (liveOn || wasmOn);
-          liveToggle.setAttribute("aria-pressed", livePressed ? "true" : "false");
-          liveToggle.disabled = !liveTransportAvailable();
-        }
-        if (wasmBtn) { wasmBtn.setAttribute("aria-pressed", wasmOn ? "true" : "false"); }
-        if (modeHint) {
-          modeHint.textContent = (liveOn || wasmOn) ? "interactive — click / scroll the preview"
-            : (liveTransportAvailable() ? "static snapshot" : "static snapshot (no live lane)");
-        }
-      }
-      if (liveToggle) {
-        liveToggle.addEventListener("click", function () {
-          if (wasmBtn) {
-            // Daemon lane specifically — the Wasm button owns the in-browser lane.
-            if (live && live.checked) { setMode("png"); }
-            else if (live && !live.disabled) { setMode("live"); }
-          } else if (anyLiveActive()) { setMode("png"); }
-          else { var m = bestLiveMode(); if (m) setMode(m); }
-        });
-      }
-      if (wasmBtn) {
-        wasmBtn.addEventListener("click", function () {
-          if (wasmActive()) { setMode("png"); } else { setMode("wasm"); }
-        });
-      }
-      if (rcBtn) {
-        rcBtn.addEventListener("click", function () {
-          if (rcActive()) { setMode("png"); } else { setMode("rc"); }
-        });
-      }
-      // ---- RC backend selector -----------------------------------------------------------------
-      // Chips choose which Remote Compose player draws the stage: `js` (the in-browser canvas lane,
-      // via setMode("rc")), or a server-side player (`java` / `cmp-android`) that re-renders the PNG
-      // with rcPlayer=<wire> (see query()). `cmp-jvm` is rendered disabled (no Skiko draw path yet).
-      // The pressed chip tracks the active backend and stays in sync with the js canvas toggle.
-      if (rcBackendsEl) {
-        var rcChips = rcBackendsEl.querySelectorAll(".cp-rc-backend[data-rc-backend]");
-        var rcDefaultBackend = rcBackendsEl.getAttribute("data-default") || "";
-        // Assign the hoisted stub with the real reconciler (see the declaration before query()).
-        syncRcBackendChips = function () {
-          // The js lane wins the "current" marker whenever its canvas is active; otherwise it's the
-          // picked server backend, or the default until the visitor picks one. Any OTHER active lane
-          // (Live / Wasm) means no server backend is on screen, so no chip is marked current.
-          var active =
-            rcActive() ? "js"
-            : anyLiveActive() ? ""
-            : (rcPlayerPicked ? rcPlayerBackend : rcDefaultBackend);
-          Array.prototype.forEach.call(rcChips, function (c) {
-            var on = c.getAttribute("data-rc-backend") === active;
-            c.setAttribute("aria-pressed", on ? "true" : "false");
-          });
-        };
-        function pickRcBackend(w) {
-          if (w === "js") {
-            // The client canvas lane. Leave the server pick untouched so returning to a server
-            // backend restores it. setMode("rc") (via enterMode) closes any Live/Wasm lane, opens the
-            // canvas, and re-syncs the chips; if the canvas is already up there's nothing to do but
-            // reconcile the marker.
-            rcPlayerPicked = false;
-            if (!rcActive()) setMode("rc");
-            else syncRcBackendChips();
-          } else {
-            // A server-side backend. Record the pick FIRST so the single static-lane render carries
-            // rcPlayer=<wire>, then transition to the static snapshot exactly once for EVERY other
-            // lane — js canvas, Live, or Wasm. setMode("png") → enterMode("png") closes them all,
-            // renders the snapshot once (no racing double render), and re-syncs the chips. Without
-            // this a pick made while Live/Wasm was active only reloaded the hidden <img> and left the
-            // interactive renderer on screen under a pressed chip.
-            rcPlayerBackend = w;
-            rcPlayerPicked = true;
-            setMode("png");
-          }
-        }
-        Array.prototype.forEach.call(rcChips, function (c) {
-          if (c.disabled) return;
-          c.addEventListener("click", function () {
-            pickRcBackend(c.getAttribute("data-rc-backend"));
-          });
-        });
-        syncRcBackendChips();
-      }
-      // Keep the live canvas overlay tracking the snapshot's slot when the page reflows (the Wasm
-      // overlay has its own resize hook below; this covers a live session with no Wasm app).
-      window.addEventListener("resize", function () {
-        if (live && live.checked && !canvas.hidden) fitLiveCanvas();
-      });
-      // Re-pin the active overlay whenever the snapshot image itself loads — its first render, or a
-      // re-render at a new size. positionOverlay/fitLiveCanvas measure `img.getBoundingClientRect()`,
-      // which only becomes final once the new bytes are decoded, so without this an overlay picked
-      // (e.g. Live selected before the first /render lands) or a re-rendered snapshot kept the stale
-      // box until the next window resize (issue #2359).
-      img.addEventListener("load", function () {
-        if (live && live.checked && !canvas.hidden) fitLiveCanvas();
-        // Mirror the Wasm resize handler: the checkerboard phase moves with the overlay box, so
-        // re-hand the patch (which carries bgPhase) to a ready app — not just reposition the frame.
-        if (wasmActive()) {
-          positionWasmFrame();
-          if (wasmReady && wasmFrame.contentWindow) {
-            wasmFrame.contentWindow.postMessage(wasmOverridePatch(), "*");
-          }
-        }
-      });
-      if (wasmToggle) {
-        // The app posts "cp-wasm-ready" once its first frame is on the canvas — the swap signal.
-        // Match on source (the known frame's contentWindow), not e.origin — robust regardless of
-        // the frame's origin, and the payload is a fixed string so there's no data surface.
-        window.addEventListener("message", function (e) {
-          if (e.source !== wasmFrame.contentWindow || e.data !== "cp-wasm-ready") return;
-          revealWasm();
-        });
-        // Fallback for an app build that predates the ready signal: reveal a beat after the
-        // document's load event rather than holding the snapshot forever.
-        wasmFrame.addEventListener("load", function () {
-          setTimeout(function () { revealWasm(); }, 8000);
-        });
-        // The overlay tracks the snapshot's box, which moves when the page reflows — and the
-        // checkerboard phase moves with it, so re-hand it to the app.
-        window.addEventListener("resize", function () {
-          if (!wasmActive()) return;
-          positionWasmFrame();
-          if (wasmReady && wasmFrame.contentWindow) {
-            wasmFrame.contentWindow.postMessage(wasmOverridePatch(), "*");
-          }
-        });
-      }
-      if (fs) {
-        fs.addEventListener("input", function () {
-          fsVal.textContent = fs.value;
-          fontScaleTouched = true;
-          onControlsChanged();
-        });
-      }
-      fields.forEach(function (f) {
-        var el = document.getElementById("cp-" + f);
-        if (el) el.addEventListener("change", onControlsChanged);
-      });
-      // Size mode: show only the input rows the chosen mode uses (Within shows both min + max), then
-      // re-render. The number inputs re-render on "input" (live typing) like the locale field.
-      var sizeMode = document.getElementById("cp-sizeMode");
-      if (sizeMode) {
-        var syncSizeRows = function () {
-          var m = sizeMode.value;
-          var show = { fixed: m === "fixed", min: m === "min" || m === "within",
-            max: m === "max" || m === "within" };
-          ["fixed", "min", "max"].forEach(function (g) {
-            var row = document.getElementById("cp-size-" + g);
-            if (row) row.hidden = !show[g];
-          });
-        };
-        syncSizeRows();
-        sizeMode.addEventListener("change", function () { syncSizeRows(); onControlsChanged(); });
-        ["cp-fixedW", "cp-fixedH", "cp-minW", "cp-minH", "cp-maxW", "cp-maxH"].forEach(function (id) {
-          var el = document.getElementById(id);
-          if (el) el.addEventListener("input", onControlsChanged);
-        });
-      }
-      // Overlay toggles are live-only: they push a fresh setOverrides through the open stream and do
-      // nothing otherwise. They get their own handler rather than onControlsChanged so a toggle mid
-      // connect (ws not yet readyState 1) can't fall through to the snapshot / wasm-auto-enable
-      // branches — an overlay never applies to a baked PNG or the in-browser tier.
-      function onOverlayChanged() {
-        if (live.checked && ws && ws.readyState === 1) {
-          ws.send(JSON.stringify({ type: "setOverrides", overrides: liveOverrides() }));
-        }
-      }
-      Array.prototype.forEach.call(overlayToggles, function (el) {
-        el.addEventListener("change", onOverlayChanged);
-      });
-      // Author-declared **named knobs** (label, count, colour, …) re-render on edit (text/number
-      // debounce via "input", toggles "change"). Unlike the app-theme selector and detected-feature
-      // toggles below, these ARE honoured by the in-browser Wasm tier (its `catalogOverride*` seeds
-      // from the `knob.<key>` patch), so a knob edit drives whichever transport is live: the Wasm
-      // iframe when it's active (or auto-enable it on a static published catalog), the daemon stream
-      // when Live is up, or a `/render` snapshot when the session can re-render.
-      function onKnobEdited() {
-        refreshLinks();
-        if (wasmActive()) {
-          if (wasmReady && wasmFrame.contentWindow) {
-            wasmFrame.contentWindow.postMessage(wasmOverridePatch(), "*");
-          } else {
-            wasmFrame.src = wasmInitialSrc();
-          }
-          return;
-        }
-        if (live.checked && ws && ws.readyState === 1) {
-          ws.send(JSON.stringify({ type: "setOverrides", overrides: liveOverrides() }));
-        } else if (canRenderOverrides) {
-          refreshSnapshot();
-        } else if (staticSnapshot && wasmToggle) {
-          // A published catalog can't re-render on the server, but its in-browser app can apply the
-          // knob — auto-enable the Wasm tier and let its load carry the edit (wasmInitialSrc bakes
-          // the patch into the fragment), mirroring the display-axis auto-enable in onControlsChanged.
-          setMode("wasm");
-        }
-      }
-      document.querySelectorAll(".cp-knob").forEach(function (el) {
-        el.addEventListener(el.type === "checkbox" ? "change" : "input", onKnobEdited);
-      });
-      // The app-theme selector and detected-feature toggles route ONLY through the server daemon —
-      // an app-declared theme provider is a server-side wrapper, and focus/gesture overlays are
-      // daemon-rendered, neither of which the in-browser tier can produce — so they use a
-      // daemon-only handler and never the wasm path.
-      function onKnobChanged() {
-        refreshLinks();
-        if (live.checked && ws && ws.readyState === 1) {
-          ws.send(JSON.stringify({ type: "setOverrides", overrides: liveOverrides() }));
-        } else if (canRenderOverrides) {
-          refreshSnapshot();
-        }
-      }
-      var themeSel = document.getElementById("cp-themeProvider");
-      if (themeSel) themeSel.addEventListener("change", onKnobChanged);
-      // Detected-feature toggles (Keyboard focus) re-render on the daemon like a knob — same routing,
-      // never the wasm auto-enable path.
-      document.querySelectorAll(".cp-feature").forEach(function (el) {
-        el.addEventListener("change", onKnobChanged);
-      });
-      // Remote Compose knobs apply live in the browser when the RC canvas lane is active
-      // (setNamed*Override + repaint on the client player, no round-trip); otherwise they route
-      // through the server daemon like the theme selector / feature toggles. A text/number edit
-      // debounces on "input", a bool toggle on "change".
-      function onRcKnobChanged() {
-        if (rcActive()) { refreshLinks(); applyRcOverrides(); return; }
-        onKnobChanged();
-      }
-      document.querySelectorAll(".cp-rc-knob").forEach(function (el) {
-        el.addEventListener(el.type === "checkbox" ? "change" : "input", onRcKnobChanged);
-      });
-      // Reconcile the control enabled-state + the toggle's initial look with the session's
-      // capabilities (matches the server-rendered markup; keeps them in sync after hydration).
-      syncServerControls();
-      updateLiveToggle();
-      refreshSnapshot();
-    })();
-    """
-      .trimIndent()
 
   /**
    * The left-hand component-nav drawer: a filterable list of the session's [siblings], each linking
@@ -5338,118 +3378,6 @@ object ServeWeb {
       .trimIndent()
   }
 
-  /**
-   * Makes each collapsible controls group (`<details class="cp-group" data-cp-group="…">`) remember
-   * its open/closed state across navigations, keyed by its stable group id in `localStorage`
-   * (`cp-grp.<id>`). The server renders the defaults (Export open, everything else collapsed); a
-   * stored preference then overrides that default so a viewer who opened, say, Overrides once keeps
-   * it open on the next preview. Runs before the other viewer scripts so the panel opens in its
-   * remembered shape immediately. Silently no-ops when `localStorage` is unavailable.
-   */
-  private fun groupStickyScript(): String =
-    """
-    (function () {
-      var groups = document.querySelectorAll("details.cp-group[data-cp-group]");
-      function key(g) { return "cp-grp." + g.getAttribute("data-cp-group"); }
-      for (var i = 0; i < groups.length; i++) {
-        (function (g) {
-          try {
-            var v = localStorage.getItem(key(g));
-            if (v === "1") g.open = true;
-            else if (v === "0") g.open = false;
-          } catch (e) {}
-          g.addEventListener("toggle", function () {
-            try { localStorage.setItem(key(g), g.open ? "1" : "0"); } catch (e) {}
-          });
-        })(groups[i]);
-      }
-    })();
-    """
-      .trimIndent()
-
-  /**
-   * Toggle wiring for the two viewer drawers. Each toggle flips an open-state class on `.cp-viewer`
-   * (`cp-controls-open` for the right overrides drawer, `cp-nav-open` for the left component nav)
-   * and mirrors it into `aria-expanded`; the drawers themselves show/hide purely via CSS on those
-   * classes. The nav's own `×` closes it, and its search box filters the component list in place.
-   */
-  private fun drawerScript(): String =
-    """
-    (function () {
-      var viewer = document.querySelector(".cp-viewer");
-      if (!viewer) return;
-      var scrim = document.getElementById("cp-scrim");
-      function isMobile() {
-        return !!(window.matchMedia && window.matchMedia("(max-width: 640px)").matches);
-      }
-      // On a phone the drawers open as bottom sheets over the preview; show a scrim behind whichever
-      // is open. On desktop they're inline columns and the scrim rule never applies.
-      function syncScrim() {
-        var anyOpen =
-          viewer.classList.contains("cp-controls-open") ||
-          viewer.classList.contains("cp-nav-open");
-        if (scrim) scrim.classList.toggle("cp-scrim-on", anyOpen);
-      }
-      function toggleId(cls) {
-        return cls === "cp-nav-open" ? "cp-nav-toggle" : "cp-controls-toggle";
-      }
-      function setOpen(cls, open) {
-        // On mobile, opening one sheet closes the other so they never stack.
-        if (open && isMobile()) {
-          var other = cls === "cp-controls-open" ? "cp-nav-open" : "cp-controls-open";
-          if (viewer.classList.contains(other)) {
-            viewer.classList.remove(other);
-            var ob = document.getElementById(toggleId(other));
-            if (ob) ob.setAttribute("aria-expanded", "false");
-          }
-        }
-        viewer.classList.toggle(cls, open);
-        var b = document.getElementById(toggleId(cls));
-        if (b) b.setAttribute("aria-expanded", open ? "true" : "false");
-        syncScrim();
-      }
-      function bindToggle(btnId, cls) {
-        var btn = document.getElementById(btnId);
-        if (!btn) return;
-        btn.addEventListener("click", function () {
-          setOpen(cls, !viewer.classList.contains(cls));
-        });
-      }
-      // Start with the overrides drawer collapsed on a phone so the preview leads; the sticky toggle
-      // bar keeps it (and the component list) one tap away as a bottom sheet.
-      if (isMobile()) setOpen("cp-controls-open", false);
-      bindToggle("cp-controls-toggle", "cp-controls-open");
-      bindToggle("cp-nav-toggle", "cp-nav-open");
-      var close = document.getElementById("cp-nav-close");
-      if (close) close.addEventListener("click", function () { setOpen("cp-nav-open", false); });
-      // Tap the scrim to dismiss whichever bottom sheet is open.
-      if (scrim)
-        scrim.addEventListener("click", function () {
-          setOpen("cp-controls-open", false);
-          setOpen("cp-nav-open", false);
-        });
-      syncScrim();
-      var search = document.getElementById("cp-nav-search");
-      if (search)
-        search.addEventListener("input", function () {
-          var query = search.value.trim().toLowerCase();
-          var items = document.querySelectorAll("#cp-nav-list .cp-nav-item");
-          var shown = 0;
-          for (var i = 0; i < items.length; i++) {
-            var el = items[i];
-            var hay = (el.getAttribute("data-search") || "").toLowerCase();
-            // The current preview (aria-current) always stays; others hide on a filter miss.
-            var keep = query === "" || el.hasAttribute("aria-current") || hay.indexOf(query) !== -1;
-            el.parentNode.hidden = !keep;
-            if (keep) shown++;
-          }
-          var empty = document.getElementById("cp-nav-empty");
-          if (empty) empty.hidden = shown > 0;
-        });
-    })();
-    """
-      .trimIndent()
-
   private fun document(
     title: String,
     body: String,
@@ -5504,7 +3432,7 @@ object ServeWeb {
         <meta charset="utf-8">
         <meta name="viewport" content="width=device-width, initial-scale=1">$unfurlBlock
         <title>${WebEscaping.htmlEscape(title)}</title>
-        <style>$STYLE</style>
+        <link rel="stylesheet" href="${assetHref("serve.css")}">
         <!-- Apply the sticky Background/Transparent choice before first paint (no checkerboard flash). -->
         <script>try{if(localStorage.getItem("cp-bg")==="off")document.documentElement.classList.add("cp-bg-transparent");}catch(e){}</script>
       </head>
